### 2018-04-05 #209 @richelbilderbeek
submission: babette
`babette` provides for a full workflow alternative of four different
GUI and command-line tools.

```
Package: babette
Title: Control BEAST2 from R
Version: 1.2.2
Author: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
Maintainer: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
Description: 'BEAST2' (<http://www.beast2.org>) is a widely used
  Bayesian phylogenetic tool, that uses DNA/RNA/protein data
  and many model priors to create a posterior of jointly estimated
  phylogenies and parameters.
  'BEAST2' is commonly accompanied by 'BEAUti 2' (<http://www.beast2.org>),
  'Tracer' <http://tree.bio.ed.ac.uk/software/tracer/>
  and Densitree (<http://www.beast2.org>).
  'babette' provides for an alternative workflow of using
  all these tools seperately. This allows doing complex Bayesian
  phylogenetics easily and reproducibly from R.
License: GPL-3 | file LICENSE
    beautier,
    beastier,
    phangorn,
    tracerer,
    testit,
    xml2
    devtools,
    lintr,
    nLTT,
    jimhester/covr,
    jimhester/lintr,
    richelbilderbeek/beautier,
    richelbilderbeek/beastier,
    richelbilderbeek/nLTT,
    richelbilderbeek/tracerer,
    MangoTheCat/goodpractice,
    KlausVigo/phangorn
```

https://github.com/richelbilderbeek/babette

[e.g., "data extraction, because the package parses a scientific data file format"]

reproducibility, as it provides a way to script three error-prone (and tedius) GUI
and one command-line tool.

Scientists that use the tool 'BEAST2' to do a Bayesian phylogenetic inference on DNA data.

No

### Deviation from this guideline

There is one point I deviate from this guideline:

When using roxygen2, add `#' @noRd` to internal functions.

There are two functions `beautier::default_params_doc` and `beastier::default_params_doc`
that are empty internal function. Their purpose is to have their parameters inherited.

Adding `#' @noRd` stops internal and exported functions to inherit these parameters.

If there is alternative, I'd be happy to hear it.

### Deviate from `goodpractice::gp()`

There is one `goodpractice::gp()` warning I ignore. About the 'Date' field in
the `DESCRIPTION` file. `goodpractice` suggest to remove it, `R CMD check --as-cran`
forces to add it. I followed the CRAN standards,

If there is way to satisfy both checks, I'd love to hear it.

Klaus Schliep, KlausVigo
Luke Harmon, lukejharmon
Jonathan Eastman, eastman
Joseph W. Brown, josephwb
