### 2018-04-12 #209 @GuangchuangYu
The example on [tracerer README](https://github.com/richelbilderbeek/tracerer) is broken since the function `get_path` was renamed to `get_tracerer_path`. Please update documentations accordingly.

The `parse_beast_trees` internally uses `rBEAST::beast2out.read.trees`, which output `multiPhylo` object without any BEAST inferences. I think you need to describe this in vignette and a link to [treeio](https://github.com/GuangchuangYu/treeio) is helpful, see also <https://github.com/ropensci/onboarding/issues/179>. `treeio` is able to parse BEAST tree with corresponding statistics.

The output of the example is `10 phylogenetic trees`:
```
trees <- parse_beast_trees(
  get_tracerer_path("beast2_example_output.trees")
)
```

while `treeio` output 11 trees.

After looking at the source code, there is an option:

> opt.burnin: how many phylogenies to discard,

which by default is setting to 0, in the function `rBEAST::beast2out.read.trees`.

Although by default it is expected to discard 0 tree, it actually discard the first one. It is a bug or intended for some reasons?

----

```r
> rBEAST::beast2out.read.trees
function (file, opt.rescale.edge.length = 1, opt.burnin = 0)
{
    tmp <- readLines(file, n = 2000, warn = FALSE)
    tmp <- which(grepl("#NEXUS", tmp))
    if (length(tmp) > 1) {
        cat(paste("\nFound #NEXUS headers, n=", length(tmp),
            ".\nDiscard all lines before last entry on line",
            tail(tmp, 1)))
        cmd <- paste("sed -i\".bak\" 1,", tail(tmp, 1) - 1, "d ",
            file, sep = "")
        system(cmd)
        cmd <- paste("sed -i\".bak2\" 1s/\\;// ", file, sep = "")
        system(cmd)
        cmd <- list.files(paste(rev(rev(strsplit(file, "/")[[1]])[-1]),
            collapse = "/"), pattern = "*bak*", full.names = TRUE)
        cat(paste("\nrm files\n", paste(cmd, collapse = "\n")))
        file.remove(cmd)
    }
    mph <- ape::read.nexus(file)
    tmp <- regexpr("[0-9]+", names(mph))
    if (any(tmp < 0))
        stop("unexpected nexus file without STATE iteration numbers")
    mph.it <- as.numeric(regmatches(names(mph), tmp))
    mph <- lapply(which(mph.it > opt.burnin), function(j) mph[[j]])
    mph.it <- mph.it[mph.it > opt.burnin]
    names(mph) <- paste("STATE_", mph.it, sep = "")
    if (opt.rescale.edge.length != 1)
        for (j in seq_along(mph)) mph[[j]]$edge.length <- mph[[j]]$edge.length *
            opt.rescale.edge.length
    mph
}
```

The implementation of `rBEAST::beast2out.read.trees` is not robust.

+ It only read first 2000 lines.
+ use system command (i.e. `sed` may not be available for Windows and may have different behaviors in Unix-like system for there are different implementation of `sed`, e.g. GNU or BSD implementation).

```r
### something like this in pure R is more favorable.

    con <- readLines(file, n = 2000, warn = FALSE)
    tmp <- which(grepl("#NEXUS", con))
    if (length(tmp) > 1) {
        cat(paste("\nFound #NEXUS headers, n=", length(tmp),
            ".\nDiscard all lines before last entry on line",
            tail(tmp, 1)))
        ##
        con <- con[tmp[length(tmp)]:length(con)]
  }
  mph <- ape::read.nexus(textConnection(con))
  ##
  ...
```
