### 2018-04-13 #209 @richelbilderbeek
Thanks for your review @GuangchuangYu

 * [ ] Update doc in README
 * [ ] switch to using treeio for reading `.trees` files
