+++ appveyor.yml
@@ -15,8 +15,18 @@ cache:
 
 # Adapt as necessary starting from here
 
+# Inspired on https://github.com/juliasilge/tidytext
+environment:
+  global:
+    WARNINGS_ARE_ERRORS: 1
+    USE_RTOOLS: true
+
 build_script:
   - travis-tool.sh install_deps
+  - ./install_beast2
+  - travis-tool.sh install_github richelbilderbeek/tracerer
+  - travis-tool.sh install_github richelbilderbeek/beautier
+  - travis-tool.sh install_github richelbilderbeek/beastier
 
 test_script:
   - travis-tool.sh run_tests
