### 2018-04-13 39f1846
tracerer reads all trees from file
