### 2018-04-14 eefba01
Install BEAST2 using beastier
