### 2018-04-14 ed6733d
beastier installs BEAST2 correctly
