### 2018-04-15 6c7bf52
AppVeyor installs BEAST2 using beastier
