### 2018-04-15 #209 @richelbilderbeek
Added AppVeyor continuous integration, used `rappdir` for using the correct folders and got everything to work under both Linux and Windows :+1:
