### 2018-04-16 5f660dc
Check RNG seed for valid values. Use beastier::install_beast2()
