+++ R/run.R
@@ -183,7 +183,7 @@ run <- function(
     }
   }
   new_names <- names(out)
-  new_names[1:length(fasta_filenames)] <- paste0(
+  new_names[seq_along(fasta_filenames)] <- paste0(
     beautier::get_ids(fasta_filenames), "_trees"
   )
   names(out) <- new_names
