### 2018-04-17 8625e4c
Fixed https://github.com/richelbilderbeek/babette/issues/10
