### 2018-04-26 #209 @noamross
It took a bit to get reviewers, @richelbilderbeek, but the search made it clear that there is a lot of interest in these packages!  Like Guangchuang, there were several folks who had short comments but were not able to do a full review.  One comment that came up is that one might want to use the **xml2** package to generate and validate XML inputs rather than manipulating strings.  I suggest that the reviewers consider this and determine whether they agree.
