### 2018-04-28 b586a0b
Install BEAST2
