+++ README.md
@@ -29,7 +29,19 @@ If you use the `devtools` R package, this is easy:
 devtools::install_github("richelbilderbeek/babette")
 ```
 
-To install BEAST2, see [how to install BEAST2](https://github.com/richelbilderbeek/beastier/blob/master/install_beast2.md).
+`babette` assumes that BEAST2 is installed. To install BEAST2, from R do:
+
+```{r}
+library(beastier)
+install_beast2()
+```
+
+This will download and extract BEAST2 to:
+
+OS|Full path
+---|---
+Linux|`~/.local/share/beast/bin/beast.jar`
+Windows|`C:/Users/<username>/Local/beast/bin/beast.jar`
 
 ## FAQ
 
