+++ doc/faq.md
@@ -78,7 +78,7 @@ Initially, the logo was a low-tech remake of Babette, a maid in Beauty and the B
 To prevent problems with Disney, a different logo was picked.
 
 The current logo shows a swan, an animal considered to be graceful.
-The swan is drawn by Jose Scholten, who kindly allowed her work to
+The swan is drawn by Jose Scholte, who kindly allowed her work to
 be used for free, by attribution.
 
 ## How did you convert the fuzzy white background to one single color?
