### 2018-04-28 #209 @richelbilderbeek
The reason of not using `xml2` is in the history of the package. It started as a hackerish thing, not intended for re-use. After a while, when `babette` was in her final phase, I have tried out `xml2` for a short while. But I could not convince myself to redo the architecture of the back-end, when everything already worked with 100% code coverage. If reviewers insist, I do have a reason to redo the back-end, but I do not see additional added value (yet?).
