### 2018-05-03 #209 @richelbilderbeek
There are some requests by reviewers of the manuscript of `babette` [1] for renaming the function `babette::run` to `babette::bbt_run`, as the latter name is less likely to conflict with other packages. I agree, but would like permission to do so while under code review. As the deadline for the manuscript resubmission is already **next Monday**, I cannot wait until the end of the code review.

@noamross : do I have permission to change this function's name while under code review? Of course, all usages of it will also be updated and tested.

## References

 * [1] Bilderbeek, Richel J.C., Etienne, Rampal S., "babette: BEAUti 2, BEAST2 and Tracer for R". bioRxiv 271866; doi: https://doi.org/10.1101/271866
