+++ R/plot_densitree.R
@@ -4,7 +4,7 @@
 #'   \link[phangorn]{densiTree} function
 #' @author Richel J.C. Bilderbeek
 #' @examples
-#'   out <- run(
+#'   out <- bbt_run(
 #'    get_babette_path("anthus_aco.fas"),
 #'    mcmc = create_mcmc(chain_length = 10000, store_every = 1000)
 #'  )
