+++ man/plot_densitree.Rd
@@ -16,7 +16,7 @@ plot_densitree(phylos, ...)
 Calls \code{phangorn}'s \link[phangorn]{densiTree} function
 }
 \examples{
-  out <- run(
+  out <- bbt_run(
    get_babette_path("anthus_aco.fas"),
    mcmc = create_mcmc(chain_length = 10000, store_every = 1000)
  )

[…]
