+++ vignettes/examples.Rmd
@@ -103,7 +103,7 @@ posterior <- bbt_run(
   "my_fasta.fas",
   mcmc = mcmc,
   mrca_priors = create_mrca_prior(
-    taxa_names = get_taxa_names("my_fasta.fas"),
+    taxa_names = sample(get_taxa_names("my_fasta.fas"), size = 3),
     alignment_id = get_alignment_id("my_fasta.fas"),
     is_monophyletic = TRUE,
     mrca_distr = create_normal_distr(
