### 2018-05-07 c6941c6
Add step-by-step vignette
