+++ vignettes/examples.R
@@ -49,7 +49,7 @@ posterior <- bbt_run(
   "my_fasta.fas",
   mcmc = mcmc,
   mrca_priors = create_mrca_prior(
-    taxa_names = get_taxa_names("my_fasta.fas"),
+    taxa_names = sample(get_taxa_names("my_fasta.fas"), size = 3),
     alignment_id = get_alignment_id("my_fasta.fas"),
     is_monophyletic = TRUE,
     mrca_distr = create_normal_distr(
@@ -80,26 +80,12 @@ ggplot2::ggplot(
 ) + ggplot2::geom_line(ggplot2::aes(y = birthRate))
 plot_densitree(posterior$my_alignment_trees, width = 2)
 
-## ----example_4, cache=TRUE-----------------------------------------------
-posterior <- bbt_run(
-  "my_alignment.fas",
-  clock_models = create_rln_clock_model(),
-  mcmc = mcmc
-)
-
 ## ----fig.width=7, fig.height=7-------------------------------------------
 ggplot2::ggplot(
   data = posterior$estimates,
   ggplot2::aes(x = Sample)
 ) + ggplot2::geom_line(ggplot2::aes(y = birthRate))
 plot_densitree(posterior$my_alignment_trees, width = 2)
-
-## ----example_5, cache=TRUE-----------------------------------------------
-posterior <- bbt_run(
-  "my_alignment.fas",
-  tree_priors = create_bd_tree_prior(), 
-  mcmc = mcmc
-)
 
 ## ----fig.width=7, fig.height=7-------------------------------------------
 ggplot2::ggplot(
@@ -108,17 +94,12 @@ ggplot2::ggplot(
 ) + ggplot2::geom_line(ggplot2::aes(y = BDBirthRate))
 plot_densitree(posterior$my_alignment_trees, width = 2)
 
-## ----example_6, cache=TRUE-----------------------------------------------
-posterior <- bbt_run(
-  "my_alignment.fas",
-  tree_priors = create_yule_tree_prior(
-    birth_rate_distr = create_normal_distr(
-      mean = create_mean_param(value = 1.0),
-      sigma = create_sigma_param(value = 0.1)
-    )
-  ),
-  mcmc = mcmc
-)
+## ----fig.width=7, fig.height=7-------------------------------------------
+ggplot2::ggplot(
+  data = posterior$estimates,
+  ggplot2::aes(x = Sample)
+) + ggplot2::geom_line(ggplot2::aes(y = birthRate))
+plot_densitree(posterior$my_alignment_trees, width = 2)
 
 ## ----fig.width=7, fig.height=7-------------------------------------------
 ggplot2::ggplot(
@@ -127,43 +108,12 @@ ggplot2::ggplot(
 ) + ggplot2::geom_line(ggplot2::aes(y = birthRate))
 plot_densitree(posterior$my_alignment_trees, width = 2)
 
-## ----example_7, cache=TRUE-----------------------------------------------
-posterior <- bbt_run(
-  "my_alignment.fas",
-  site_models = create_hky_site_model(
-    gamma_site_model = create_gamma_site_model(prop_invariant = 0.5)
-  ),
-  mcmc = mcmc
-)
-
 ## ----fig.width=7, fig.height=7-------------------------------------------
 ggplot2::ggplot(
   data = posterior$estimates,
   ggplot2::aes(x = Sample)
 ) + ggplot2::geom_line(ggplot2::aes(y = birthRate))
 plot_densitree(posterior$my_alignment_trees, width = 2)
-
-## ----example_8, cache=TRUE-----------------------------------------------
-posterior <- bbt_run(
-  "my_alignment.fas",
[…]
