+++ NEWS.md
@@ -2,6 +2,24 @@
 
 Newest versions at top.
 
+## babette 1.2.4 (2018-05-17)
+
+### NEW FEATURES
+
+  * None
+
+### MINOR IMPROVEMENTS
+
+  * Tagged for [the academic article about babette](https://github.com/richelbilderbeek/babette_article)
+
+### BUG FIXES
+
+  * None
+
+### DEPRECATED AND DEFUNCT
+
+  * None
+
 ## babette 1.2.3 (2018-05-04)
 
 ### NEW FEATURES
