+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: babette
 Title: Control BEAST2 from R
-Version: 1.2.2
+Version: 1.2.4
 Author: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
 Maintainer: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
 Description: 'BEAST2' (<http://www.beast2.org>) is a widely used
