+++ NEWS.md
@@ -55,3 +55,4 @@ Newest versions at top.
 ### DEPRECATED AND DEFUNCT
 
   * None
+
