### 2018-05-17 #209 @bjoelle
Estimated hours spent reviewing: 25-30

---

This package and its dependencies beautier, beastier and tracerer provide a R interface for setting up a BEAST2 analysis, running it and processing the results. I think this interface is a good alternative to the existing graphical interfaces, especially BEAUti which requires a lot of manipulations to set up even a simple analysis. In particular these packages will be very useful for people who want to integrate BEAST2 in an existing R pipeline.

Overall I think the packages are well done, although the documentation is lacking in some places. All the packages are missing package documentation (required by rOpenSci guidelines).
Detailed comments are below, sorted by the package to which they apply. Please also check `devtools::spell_check`, which shows some typos in documentation in all packages.

Regarding the suggestion about using xml2 instead of strings, the only real advantage I see to it is making it possible to load BEAST2 configuration files to modify or add an element instead of having to rebuild the model in babette, which might help/encourage people to transition to using babette. I am happy to leave the decision to the author on that.

The installation worked fine on Mac and Windows.

#### babette
- this package needs a clearer statement of need (as per rOpenSci guidelines): some of the text from the onboarding issue and the description would fit nicely.
- `bbt_run`: the @return documentation is missing.
- `bbt_run`: The default behaviour is to place the output in temporary files which are deleted at the end of the run. I would add a warning about that in the doc or in the code, as I don't think this would be expected by most users.
- `bbt_run`: the doc should mention that rng_seed can be NA and what happens in that case (i.e the seed will be set to the BEAST2 default 127).

- `bbt_run.R`: typos l119 "Must have as many FASTA filenames as BEAST2 output trees filenames"

#### beautier

##### Structure
- The structure is my main concern with this package, as it contains a lot of code duplication and many functions need to run through all possible types of a certain object to figure out what to do with it. I think this makes the package hard to maintain and to extend to additional models.
The main issue in my opinion is that the objects (site models, clock models, etc) do not store their dependencies in a generic way, so that there is no way to know what objects a model depends on unless you know which specific model it is. My suggestion would be to add a list of components to all objects - for instance a TN93 site model would have `components = c("kappa_1", "kappa_2")`, a HKY would have `components = "kappa"` and so on. This would allow you to save a lot of code in functions like `init_site_models`, which could do things such as `for(x in model$components) model[[paste0(x,"_param")]] = init_param(model[[paste0(x,"_param")]])` instead of needing one subfunction per type of model.
Separating the different parameters (alpha, beta, gamma, etc) also leads to a lot of code duplication. It is useful for creating parameters as their default priors are different, but I would remove it for the other functions: the many functions `is_%sth%_param(x)` can be replaced by one function `is_named_param(x, name)` and the `parameter_to_xml` functions also have a lot of redundancy.

- I found it confusing that the code relative to the interaction of MRCA priors and clock models is in a different place for strict and relaxed clocks, and it seems to have led to different behaviour between the 2 clocks (see comments in Functionality section). I would recommend putting everything in the `mrca_prior_to_xml` functions.
- `get_id`, `get_ids`: These functions appear to be duplicated by `get_alignment_id` and `get_alignment_ids`.
- I think this package contains too many files, many of which contain only one small function. Gathering functions which are strongly linked to each other (for instance, `are_site_models`, `is_site_model`, `is_site_model_name` and `get_site_model_names`) in one file would make the package easier to read and search in my opinion.

##### Functionality
- the relaxed clock model doesn't work with MRCA priors that have a time calibration distribution: the state, operators, distribution and tracelog sections are all missing elements compared to an XML produced with BEAUti. This issue appears to be linked to `clock_models_to_xml_state` l24-25: the comment is unclear as to whether it is supposed to be checking for any or no MRCA priors, and no check is performed here for either.
- `mrca_prior_to_xml_tracelog`, `mrca_prior_to_xml_operators`, `mrca_prior_to_xml_prior_distr`, `mrca_prior_to_xml_state`: these functions only add the parts relative to the strict clock if the MRCA prior is monophyletic, whereas in BEAUTi these parts are added if the MRCA prior has a distribution associated with it, whether it is monophyletic or not.

- the distribution hyper parameters (alpha, beta, mean etc - all except lambda) can be set with `estimate = TRUE`, but the xml will not be created correctly with those settings (missing distribution, operators and state sections). I think the code should not allow this setting until this functionality is fully implemented.
- creating a TN93 model with the default settings creates two log-normal distributions as priors for kappa1 and kappa2 where the S parameter has `lower=0`, `upper=0` and `value=1.25`. This works if S is not estimated, because BEAST2 only checks upper and lower when modifying the value (not when initialising it) but it will break if S is estimated.

- `get_phylo_crown_age` will not work with sampling through time/non-ultrametric trees. I think these are not supported at the moment anyway, but it might be good to keep it in mind.
- `is_index_of_first_shared_clock_model`: this function will return TRUE if the tested model is the 2nd of 3 shared clock models, in contradiction with the doc. It will also always return TRUE if the clock model at index i has `id=NA`.
- `is_distr`: this function just checks if the input has a name attribute, which doesn't match with the name or documentation of the function
- `get_site_model_n_params` and `get_site_model_n_distrs` do not check if the site model has gamma rate categories, they just assume it does

##### Documentation
- `kappa_1`, `kappa_2` and the tree prior parameters are always estimated and this cannot be changed - I think this should be mentioned in the corresponding doc.
- the doc should make it clearer that the function used to build the branchRateModel section is `clock_model_to_xml_lh_distr` if there are no MRCA priors, and `mrca_prior_to_xml_lh_distr` if there are MRCA priors, and that `mrca_prior_to_xml_lh_distr` only supports strict clocks at the moment
- `fasta_file_to_sequences`: both the documentation and the input variable (`fasta_filenames`) wrongly indicate that this function can take multiple files
- `get_id`, `get_ids`: I don't understand what "conclude the ID" means in this doc. I would replace that with "extract the file names without extensions" or something similar.

##### Minor comments
- all the functions `clock_models_to_xml` start by calling `get_unlinked_clock_models`. This call could be placed earlier (`create_beast2_input_run` is my suggestion) which would save the repetition.
- there is an empty file `to_do.R` in the package
- `fastas_to_phylos`: the main description and @return say this creates a random phylogeny instead of one or more depending on the input
- `is_index_of_non_first_shared_clock_model`, `create_beast2_input_init`: wrong doc
- `get_site_model_n_params`: the doc refers to number of distributions instead of parameters.
- `is_in_patterns`: mismatched doc ("Create a random beta distribution")
- `remove_first_brm_clock_rate`: missing doc
- `remove_multiline`: consequetive -> consecutive
- `clock_model_to_xml_tracelog`: the seealso links to the same function
- `has_xml_closing_tag`: opening -> closing
- `init_gamma_site_model`: the doc refers to "all site models" but the function takes only one as input
- `mrca_prior_to_xml_taxonset`, `mrca_priors_to_xml_prior_distr`: the top description of these functions is confusing
- `has_xml_short_closing_tag`: at the end of the text -> in one of the lines of text
- `create_beast2_input_run`: state section -> run section

#### beastier
- `run_beast2`: I think that this function should give a warning before erasing `beast_log_filename` and `beast_trees_filenames` if they exist, because I wouldn't have expected those files to be touched at all if I have specified different values for `output_log_filename` and `output_trees_filenames`. A better solution would be to move/rename these files temporarily if they already exist, and restore them once the run is done.
- `run_beast2`, `create_beast2_run_cmd`: the -overwrite option for BEAST2 is not for overwriting the state file (it will always be overwritten), it is for overwriting the log & trees files. Thus at the moment setting overwrite_state_file does nothing (it's also always set back to TRUE at l107 in `run_beast2`).
- `run_beast2`, `create_beast2_run_cmd`: the doc should mention that rng_seed can be NA and what happens in that case (i.e the seed will be set to the BEAST2 default 127). There is also no way of setting the seed to random at the moment, so doing that if nothing is specified would be a better choice in my opinion.
- `gives_beast2_warning`: This function has a lot of code in common with `is_beast2_input_file`, so I would suggest putting the common code in a separate function that would be called by both.

- `has_unique_ids`: the doc of this function implies that it checks both parameter and distribution ids, but `get_duplicate_ids` only checks real parameters
- `is_beast2_input_file`: l38-40 should be removed - they don't do anything and status_code is 0 even if the file is invalid

- `create_posterior`: it's unclear what this function is meant to be used for, as it's not exported or called from anywhere else in the package
- `gives_beast2_warning`: this doc should link to `is_beast2_input_file` rather than `are_beast2_input_lines` since the input is a file in both.

#### tracerer
- `calc_stderr_mean`: this function asks for the sample_interval as input but doesn't seem to use it. To simplify the code I would also suggest using this formula instead of the C code: `stderr_mean = stats::sd(trace)/sqrt(length(trace))`.
- `parse_beast_output_files`, `parse_beast_posterior`: these functions need more detailed @return documentation
- `calc_summary_stats.R`: the doc in this entire file refers to calculating ESS but this function calculates other things too: a list of all the stats calculated should be included in the @return doc. The @param traces doc for `calc_summary_stats` is also confusing ("a numeric vector or data frame or numeric vector")
