### 2018-05-19 #209 @richelbilderbeek
I am impressed by @bjoelle's review, thanks so much! I will happily implement that feedback. Sure, I will wait after the other reviewer is finished, to do so in one go.
