### 2018-05-28 de94d32
Use temporary files in test
