### 2018-05-28 f2791be
@lintr-bot
