+++ tests/testthat/test-bbt_run.R
@@ -53,9 +53,11 @@ test_that("use, one alignment, verbose, cleanup", {
 test_that("use, one alignment, verbose, no cleanup", {
 
   fasta_filenames <- get_babette_path("anthus_aco.fas")
-  beast2_input_filename <-tempfile(fileext = ".xml")
+  beast2_input_filename <- tempfile(fileext = ".xml")
   beast2_output_log_filename <- tempfile(fileext = ".log")
-  beast2_output_trees_filenames <- tempfile(pattern = beautier::get_ids(fasta_filenames), fileext = ".xml")
+  beast2_output_trees_filenames <- tempfile(
+    pattern = beautier::get_ids(fasta_filenames), fileext = ".xml"
+  )
   beast2_output_state_filename <- tempfile(fileext = ".xml.state")
 
   testit::assert(!file.exists(beast2_input_filename))
