+++ R/babette.R
@@ -1,0 +1,18 @@
+#' babette: A package for Bayesian phylogenetics.
+#'
+#' 'babette' provides for an alternative workflow of using
+#' the popular phylogenetics tool BEAST2, including
+#' it peripheral tools. From an alignment and
+#' prior selection, a posterior of jointly estimated
+#' phylogenies and parameter estimates is generated.
+#'
+#' @section babette subpackages:
+#'
+#' 'babette' is mostly a glue between three subpackages:
+#'   * \link[beautier](beautier)
+#'   * beastier
+#'   * tracerer
+#'
+#' @docType package
+#' @name babette
+NULL
