### 2018-06-01 17ebe33
Better package description
