### 2018-06-01 b1c8dc0
Fix spelling errors, thanks @bjoelle
