+++ R/get_path.R
@@ -1,6 +1,6 @@
-#' Get the full path of a file in the 'inst/extdata' folder
+#' Get the full path of a file in the \code{inst/extdata} folder
 #' @param filename the file's name, without the path
-#' @return the filename's full path
+#' @return the full path of the filename
 #' @author Richel J.C. Bilderbeek
 #' @seealso for more files, use \code{\link{get_babette_paths}}
 #' @examples
