### 2018-06-01 8bcef89
Add 'hunspell' package to .travis.yml
