+++ .travis.yml
@@ -4,6 +4,7 @@ cache: packages
 r_packages:
   - devtools
   - ggplot2
+  - hunspell
   - xml2
 
 r_github_packages:
