### 2018-06-02 9810203
Use ropensci/spelling
