+++ DESCRIPTION
@@ -27,6 +27,7 @@ Imports:
     testit,
     xml2
 Suggests:
+    spelling,
     devtools,
     ggplot2,
     hunspell,
@@ -44,3 +45,4 @@ Remotes:
     richelbilderbeek/tracerer,
     MangoTheCat/goodpractice,
     KlausVigo/phangorn
+Language: en-US
