+++ .travis.yml
@@ -15,6 +15,7 @@ r_github_packages:
   - richelbilderbeek/tracerer
   - MangoTheCat/goodpractice
   - KlausVigo/phangorn
+  - ropensci/spelling
 
 before_script:
   - sudo apt-get install -qq oracle-java8-installer # Java 8
