### 2018-06-02 6f5a958
Strict spell checking
