+++ tests/spelling.R
@@ -1,1 +1,1 @@
-spelling::spell_check_test(vignettes = TRUE, error = FALSE)
+spelling::spell_check_test(vignettes = TRUE, error = TRUE)
