### 2018-06-02 3ac2d03
Add @return
