+++ R/bbt_run.R
@@ -25,7 +25,27 @@
 #' @param beast2_jar_path path to the BEAST2 jar file
 #' @param verbose set to TRUE for more output
 #' @param cleanup set to FALSE to keep all temporary files
-#' @author Richel J.C. Bilderbeek
+#' @return a list with the following elements:\cr
+#' \itemize{
+#'   \item{
+#'     \code{estimates}: a data frame with BEAST2
+#'     parameter estimates
+#'   }
+#'   \item{
+#'     \code{[alignment_id]_trees}: a \code{multiPhylo}
+#'     containing the phylogenies
+#'     in the BEAST2 posterior. \code{[alignment_id]}
+#'     is the ID of the alignment. For example,
+#'     when running \code{bbt_run} with
+#'     \code{anthus_aco.fas}, this element will have
+#'     name \code{anthus_aco_trees}
+#'   }
+#'   \item{
+#'     \code{operators}: a data frame with the
+#'     BEAST2 MCMC operator acceptances
+#'   }
+#' }
+#' @author Richèl J.C. Bilderbeek
 #' @examples
 #'  # One alignment
 #'  out <- bbt_run(
