### 2018-06-02 8e0b380
Add @return doc
