### 2018-06-02 4f6bc17
Mention rng_seed of NA in doc, thanks @bjoelle
