### 2018-06-02 1c8825a
Add 'Authors@R'
