### 2018-06-02 c264a8b
No non-ASCII
