### 2018-06-02 #209 @bjoelle
@richelbilderbeek By the way, I just realised I made a mistake in my review - if the seed is NA the BEAST seed will be random, not fixed. (I had modified my version of BEAST to fix the seed and then forgot about it when testing the package). I've updated the review to reflect that. Sorry !
