### 2018-06-02 #209 @richelbilderbeek
@bjoelle: when adding your suggestion to the documentation, I checked and concluded the same. Thanks!
