### 2018-06-15 752379e
Update citation
