### 2018-06-25 #209 @dwinter
Estimated hours spent reviewing: 17hrs

---

I must start by apologizing to @richelbilderbeek and the `ropensci` editorial team for the time I have taken to complete this review. I underestimated both my own workload and this time required to review a large an integrated project such as this. My review will not be as detailed @bjoelle's, but I have now had a chance to work with the packages and read the code and documentation.

In general, `babette` is a very useful suite of R packages that will help researchers take reproducible approaches to estimating phylogenies from sequence data. There is a real need for this package. Because key parameters for a given `BEAST2` run are generally set through a GUI program, it has been difficult to develop reproducible workflows for this very popular software.Although the need for these packages is very real, the README file on the `babette` does not currently explain this. A clear statement of purpose (similar to the one given in the submission post) would be helpful.

For the most part, the code in each package is well-written, appropriately modular meets the rOpenSci guidelines. So I think the packages will be very useful for researchers and a good addition to the rOpenSci project.

I do have some concerns about the package, namely

(a) That the code used to specify run-parameters in `beautier` is complex and may prove difficult to maintain or extend as a result.
(b) The user-exposed functions in this package are all quite "low-level", creating a daunting interface for setting prior distributions and preparing files for analysis.

This latter concern may be partly addressed by providing a more "narrative" vignette to supplment the examples in exiting ones. I provide more detail on these concerns and point out minor problems for each package  below.

#### babette

I was surprised that the "core" function `bbt_run` defaults to not saving the `.log` and `.trees` files created by a BEAST run to disk. For many users,these will be the key artefacts of a BEAST run, and important data to go into other analysis pipelines. If this behavior remains the default then it should be clearly documented in the Description field of the function-level documentation. At present, to save the trees and logs to file a user has to set both  `beast2_output_log_filename` and (possibly multiple) `beast2_output_trees_filenames`.  Consider changing this to  allow users to set a "stem" for each output file, appending  `.log` or `_{tree_id}.trees` to each particular file.

The default value for `rng_seed` in `bbt_run` set to `1`, this default is not pointed out in the package documentation. This leads to the unexpected result that two apparently-independent mcmc chains from the same dataset will be identical if the  default values are used. I suggest setting the default to `NA` (allowing BEAST to pick its own value or some random number produced in R.

The examples demonstrated in the vignette are nice, but (understandably) they are all much more simple than what I'd think of as a typical use-case. I think the package would really benefit from a "real world" example of a typical usage. Imagine a user with, say, 20 taxa, 3 loci and time calibrations for three nodes. How would they go about analysing their data and recoding their results with these package?At the moment, I'm not sure a user would find it easy to set this analysis up.

A vignette that included the narrative explaining the steps required to assign taxon-sets for the MRCA, specify the node-calibration (perhaps using `qnorm` to show the implied 95% CI)and set up the corresponding clock model would go a long way to making the package more user-friendly. (I appreciate this may be a considerable undertaking, but in the very least it should be a goal to aim for).

#### minor comments about babette

* The location of the BEAST2 jar file defaults to a directory under `rappdirs::user_data_dir()` (i.e. a system-specific data directory). Many users will have BEAST2 installed somewhere else on their system. It would be useful if users could set an environmental variable for this location so they didn't have to pass this to every call of bbt_run.

* The help for `plot_densitree` should link to the phangorn::densiTree help so users can see the un-named arguments that can be passed along.

* Many of the vignettes move files into the working directory in order to carry out an analysis, before deleting them at the end. If execution fails this can leave files in the users working directory -- something that once got one of my packages temporarily deleted from CRAN! I think best practice is to save any files required for the vignettes in `inst/exdata` and use `system.file` to define paths to those files.

#### beautier

This package attempts to cover the _very large_ set of parameter values (and their permutations) that can be set for a BEAST2 run. As one might expect, it is not capable of specifying all possible BEAST2 runs, and I appreciate the clear statement about functionality that is currently missing in the package's README. I did not a number of missing use-cases that are not currently mentioned in the  README.

* As far as I can tell it is not possible to estimate one tree from multiple alignments. (i.e. one tree prior, multiple alignments possibly each with its own site model). I suspect this is a common use-case (either specified this way or through a concatenated alignment with partitions). If I have this wrong then a partitioned dataset should be added to the examples vignette. If this is not currently possible it should be a priority for future development

* I don't think it possible to set a shifted gamma distribution (i.e. a gamma distribution with an additional "offset" value) yet. I didn't check others, but these shifted distributions are often helpful for setting node dates

Both the user-interface and the code for specifying distributions is very complex, and I am concerned about how easily this package will be to maintain. @bjoelle has made some suggestions about a more modular approach to the internal code and possibly grouping like-functions into files. I think this would be a good idea, I also wonder if you could also consider a more high-level approach to specifying these models in user-exposed code. For instance, how much power/flexibility would be lost by losing the `create_*_param()` functions and instead doing something like this:

```r
node_calibration <- create_gamma_distr(alpha=1, beta=10, estimate=FALSE)
```

I appreciate there may be reasons for taking the approach you have, but the current does make for a complex API. Whatever the API, some narrative examples
of setting up these paramaters in the documentation would be helpful.

#### Minor comments about beautier

The function `create_beast2_input` sets the user's global `scipen` option. It's not clear from the code exactly where this value is needed, but it would seem a better idea to either store the existing value (`old_scipen <- getOption("scipen")`) then ensure the value is reset using `on.exit`.

#### Tracerer

It's a shame traceR was taken as a name on CRAN! (This name is fine, given the shorter one is not available). I only have minor comments about this one.

#### Minor comments of tracerer

In the package-level documentation  consider referring the user  to `coda`, a very widely-used package for MCMC analysis that includes diagnostics and plots that are not implemented here.

In `calc_summary_stats` the documentation for `sample_interval` make it clear this is the "sampling interval used in in the MCMC run" or similar (i.e. the `sample_interval` should match the run that produced the data, and is not used to further thin output in this function).

Consider exposing `calc_hpd_interval` to users. I can imagine users may sometimes one an 80%  or 99% HPD for a particular value. At present
calc_summary_stats hard-codes a 95% HPD.

#### Beastier

I think many of my comments about `bbt_run` refer to commands that are passed on to  functions in this package.  I only have one comment specific to this package.

`install_beast2` hard-codes the URL for particular release. Consider making the release tag a user-settable argument. Perhaps include a
link to the BEAST2 home page which has more installation instructions for users.
