+++ R/bbt_run.R
@@ -14,25 +14,35 @@
 #' @param posterior_crown_age the posterior's crown age. Use NA to let
 #'   BEAST2 estimate this parameter. Use a positive value to fix the
 #'   crown age to that value
-#' @param beast2_input_filename name of the BEAST2 configuration file
+#' @param beast2_input_filename path of the BEAST2 configuration file.
+#'   By default, this file is put in a temporary folder with a random
+#'   filename, as the user needs not read it: it is used as input of BEAST2.
+#'   Specifying a \code{beast2_input_filename} allows to store that file
+#'   in a more permanently stored location.
 #' @param rng_seed the random number generator seed. Must be either
 #'   \code{NA} or a positive non-zero value. An RNG seed of \code{NA}
 #'   results in BEAST2 picking a random seed.
 #' @param beast2_output_log_filename name of the log file created by BEAST2,
-#'   containing the parameter estimates in time. By default, this temporary
-#'   file is put a default temporary folder and will be cleaned up by the
-#'   operating system. This file is a temporary file, as its content
+#'   containing the parameter estimates in time. By default, this
+#'   file is put a temporary folder with a random
+#'   filename, as the user needs not read it: its content
 #'   is parsed and returned by this function.
+#'   Specifying a \code{beast2_output_log_filename} allows to store that file
+#'   in a more permanently stored location.
 #' @param beast2_output_trees_filenames name of the one or more trees
-#'   files created by BEAST2, one per alignment. By default, these temporary
-#'   files are put a default temporary folder and will be cleaned up by the
-#'   operating system. These files are temporary files, as their content
+#'   files created by BEAST2, one per alignment. By default, these
+#'   files are put a temporary folder with a random
+#'   filename, as the user needs not read it: their content
 #'   is parsed and returned by this function.
+#'   Specifying \code{beast2_output_trees_filenames} allows to store these
+#'   one or more files in a more permanently stored location.
 #' @param beast2_output_state_filename name of the final state file created
-#'   by BEAST2, containing the operator acceptances. By default, this temporary
-#'   file is put a default temporary folder and will be cleaned up by the
-#'   operating system. This file is a temporary file, as its content
+#'   by BEAST2, containing the operator acceptances. By default, this
+#'   file is put a temporary folder with a random
+#'   filename, as the user needs not read it: its content
 #'   is parsed and returned by this function.
+#'   Specifying a \code{beast2_output_state_filename} allows to store that file
+#'   in a more permanently stored location.
 #' @param beast2_jar_path path to the BEAST2 jar file
 #' @param verbose set to TRUE for more output
 #' @param cleanup set to FALSE to keep all temporary files
