+++ vignettes/tutorial.Rmd
@@ -1,0 +1,376 @@
+---
+title: "babette Tutorial"
+author: "Richel J.C. Bilderbeek"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{babette Tutorial}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+This vignette is a tutorial how to use `babette` and its most 
+important `bbt_run` function.
+
+First, load `babette`:
+
+```{r}
+library(babette)
+```
+
+The main function of `babette` is `bbt_run`. Here is part of its help:
+
+```
+Do a full run: create a BEAST2 configuration file (like BEAUti 2), run BEAST2, parse results (like Tracer)
+
+Usage
+
+bbt_run(
+  fasta_filenames,
+  site_models,
+  clock_models,
+  tree_priors,
+  mrca_priors,
+  posterior_crown_age,
+  beast2_output_log_filename,
+  beast2_output_trees_filenames,
+  beast2_output_state_filename,
+  verbose, 
+  cleanup
+)
+```
+
+Simplifying this to all arguments that do not have a default:
+
+```
+bbt_run(
+  fasta_filenames
+)
+```
+
+## `fasta_filenames`
+
+`fasta_filenames` is the `bbt_run` argument to specify the one or more
+FASTA files to work on. `babette` is bundled with some FASTA files,
+so obtaining a path to a FASTA file is easy:
+
+```{r}
+fasta_filename <- get_babette_path("anthus_aco.fas")
+testit::assert(file.exists(fasta_filename))
+```
+
+With `fasta_filename` available, we have the minimal
+requirements to call `bbt_run` like this:
+
+```
+out <- bbt_run(fasta_filenames = fasta_filename)
+```
+
+Note that this code is not ran, as it would take too long.
+The reason this would take too long, is that 
+the MCMC run that will be executed is set to one million states by default.
+To specify the MCMC options and shorten this run, 
+the `mcmc` argument is used.
+
+## `mcmc` 
+
+`mcmc` is the `bbt_run` argument to specify the MCMC run options:
+
+```{r}
+mcmc <- create_mcmc(chain_length = 2000, store_every = 1000)
+```
+
+With these MCMC options, we can now call `bbt_run` in way that
+it will finish fast:
+
+```{r cache=TRUE}
+out <- bbt_run(
+ fasta_filenames = fasta_filename,
+ mcmc = mcmc
+)
+```
+
+The return value, `out` contains the results of the MCMC run.
+For this tutorial, visualizing `out` is ignored, as the 'Demo' vignette
+discusses this. Instead, we will work through the other `bbt_run` parameters.
+
+## `site_models` 
+
+`site_models` is the `bbt_run` argument to select one or more site models.
+As this tutorial works on a DNA alignment, such a site model can also
+be called a nucleotide substitution model.
+
+Picking a site model is easy: just type:
+
+```
+create_site_model_
+```
+
+This will trigger autocomplete to show all site models. 
+
+The simplest site model is the Jukes-Cantor DNA substutition model.
+To use this model in `babette`, do:
+
+```{r}
+site_model <- create_site_model_jc69()
+site_model <- create_jc69_site_model()
+```
+
+Using this site model:
+
+```{r cache=TRUE}
+out <- bbt_run(
+  fasta_filenames = fasta_filename,
+  site_models = site_model,
+  mcmc = mcmc
+)
+```
+
[…]

[…]
