### 2018-07-07 #209 @richelbilderbeek
Big package, big reviews, then ... well ... big number changes. Processed [the feedback specific to babette](https://github.com/richelbilderbeek/babette/issues/9), still working on its sub-packages. Will post per sub-package. Unless someone has a better idea?
