### 2018-07-07 553fca0
Fix spelling errors
