### 2018-07-07 ca6908d
Thank rOpenSci reviewers, fix #14
