+++ DESCRIPTION
@@ -3,7 +3,8 @@ Title: Control BEAST2 from R
 Version: 1.2.4
 Authors@R: c(
   person("Richel", "Bilderbeek", email = "richel@richelbilderbeek.nl", role = c("aut", "cre"), comment = c(ORCID = "0000-0003-1107-7049")),
-  person("Joelle", "Barido-Sottani", role = "rev", comment = "Joelle reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/209"))
+  person("Joelle", "Barido-Sottani", role = "rev", comment = "Joelle reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/209"),
+  person("David", "Winter", role = "rev", comment = "David reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/209"))
 Author: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
 Maintainer: Richel J.C. Bilderbeek <richel@richelbilderbeek.nl>
 Description: 'BEAST2' (<http://www.beast2.org>) is a widely used
