### 2018-07-21 #209 @richelbilderbeek
[Implemented all tracerer feedback](https://github.com/richelbilderbeek/tracerer/projects/1)
