### 2018-08-05 #209 @richelbilderbeek
[All beastier feedback has been addressed](https://github.com/richelbilderbeek/beastier/projects). Only beautier (the biggest!) to go... :rainbow:
