+++ doc/examples.md
@@ -1,5 +1,8 @@
 # Examples
 
+ * Want something more interactive? [beastier](https://github.com/richelbilderbeek/lumier) 
+   is an R Shiny app to help create the R function call needed!
+
 For all examples, do load `babette`:
 
 ```
