### 2018-08-14 7384105
Use 'beautier::get_alignment_id' instead of the obsolete and ambiguous 'beautier::get_id'
