### 2018-08-14 95f9ec9
Add 'bbt_run' @usage in documentation
