### 2018-09-11 7bf8edf
Show 'mauricer' and 'lumier' build statuses [skip ci]
