### 2018-09-11 d7a4723
Improve documentation [skip ci]
