+++ doc/README.md
@@ -1,0 +1,12 @@
+# `doc`
+
+The `babette` documentation.
+
+## Create the dependency graph from the `.dot` file
+
+```
+dot -Tps dependencies.dot -o dependencies.ps
+convert dependencies.ps dependencies.png
+```
+
+
