### 2018-09-12 aece9c8
Add Nested Sampling, from Maturana, P., Brewer, B. J., Klaere, S., & Bouckaert, R. (2017). Model selection and parameter inference in phylogenetics using Nested Sampling. arXiv preprint arXiv:1703.05471.
