### 2018-09-12 4bd9ddc
Use 'beast2_path' instead of 'beast2_jar_path'
