### 2018-09-12 1e4a958
Vignette 'Nested Sampling' works locally
