+++ vignettes/nested_sampling.Rmd
@@ -2,7 +2,7 @@
 title: "Nested Sampling"
 author: "Richel J.C. Bilderbeek"
 date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
+output: rmarkdown::pdf_document
 vignette: >
   %\VignetteIndexEntry{Nested Sampling}
   %\VignetteEngine{knitr::rmarkdown}
@@ -46,22 +46,28 @@ if (!mauricer::mrc_is_installed("NS"))
 ## Do the run
 
 ```{r}
+beast2_input_file <- tempfile(fileext = ".xml")
 beautier::create_beast2_input_file(
   input_filenames = get_babette_path("anthus_aco_sub.fas"),
-  output_filename = "~/out.xml",
+  output_filename = beast2_input_file,
   mcmc = create_mcmc_nested_sampling(chain_length = 10000)
 )
 ```
 
 ```{r}
-# out <- bbt_run(
-#   fasta_filenames = get_babette_path("anthus_aco_sub.fas"),
-#   mcmc = create_mcmc_nested_sampling(chain_length = 10000),
-#   beast2_input_filename = "~/out.xml",
-#   beast2_path = beastier::get_default_beast2_bin_path() 
-# )
+out <- bbt_run(
+  fasta_filenames = get_babette_path("anthus_aco_sub.fas"),
+  mcmc = create_mcmc_nested_sampling(chain_length = 20000, store_every = 1000),
+  beast2_input_filename = beast2_input_file,
+  beast2_path = beastier::get_default_beast2_bin_path()
+)
 ```
 
+Show the results, where `NSlikelihood` is the Nested-Sampling likelihood:
+
+```{r}
+knitr::kable(out$estimates)
+```
 
 ## References
 
