+++ DESCRIPTION
@@ -51,3 +51,4 @@ Remotes:
     MangoTheCat/goodpractice,
     KlausVigo/phangorn
 Language: en-US
+Encoding: UTF-8
