### 2018-09-13 ce2f783
Check if the needed BEAST2 packages are installed
