+++ tests/testthat/test-bbt_check_beast2_packages.R
@@ -1,0 +1,36 @@
+context("bbt_check_beast2_packages")
+
+test_that("use", {
+
+  beast2_package_name <- "NS"
+  # Check to need to install NS later
+  was_ns_installed <- mauricer::mrc_is_installed(beast2_package_name)
+
+  if (mauricer::mrc_is_installed(beast2_package_name)) {
+    mauricer::mrc_uninstall(beast2_package_name)
+  }
+  testit::assert(!mauricer::mrc_is_installed(beast2_package_name))
+
+  expect_error(
+    babette:::bbt_check_beast2_packages(
+      mcmc = create_mcmc_nested_sampling(),
+      beast2_path = get_default_beast2_bin_path()
+    ),
+    paste0(
+      "Must install 'NS' to use 'create_mcmc_nested_sampling'."
+    )
+  )
+
+  mauricer::mrc_install(beast2_package_name)
+
+  expect_silent(
+    babette:::bbt_check_beast2_packages(
+      mcmc = create_mcmc_nested_sampling(),
+      beast2_path = get_default_beast2_bin_path()
+    )
+  )
+
+  if (!was_ns_installed) {
+    mauricer::mrc_uninstall(beast2_package_name)
+  }
+})
