+++ man/get_babette_path.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_path.R
+% Please edit documentation in R/bbt_get_path.R
 \name{get_babette_path}
 \alias{get_babette_path}
 \title{Get the full path of a file in the \code{inst/extdata} folder}
