+++ man/get_babette_paths.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_paths.R
+% Please edit documentation in R/bbt_get_paths.R
 \name{get_babette_paths}
 \alias{get_babette_paths}
 \title{Get the full paths of files in the 'inst/extdata' folder}
