+++ man/babette.Rd
@@ -24,6 +24,10 @@ These are packages associated with 'beautier':
     BEAST2 from R
   }
   \item{
+    The package 'mauricer' can install
+    BEAST2 packages from R
+  }
+  \item{
     The package 'tracerer' can parse
     BEAST2 output files from R
   }
