### 2018-09-13 1a5b9e1
Update and ordered FAQ [skip ci]
