### 2018-09-13 8607557
Run Nested Samping vignette with a short run
