+++ vignettes/nested_sampling.Rmd
@@ -46,19 +46,14 @@ if (!mauricer::mrc_is_installed("NS"))
 ## Do the run
 
 ```{r}
-beast2_input_file <- tempfile(fileext = ".xml")
-beautier::create_beast2_input_file(
-  input_filenames = get_babette_path("anthus_aco_sub.fas"),
-  output_filename = beast2_input_file,
-  mcmc = create_mcmc_nested_sampling(chain_length = 10000)
-)
-```
-
-```{r}
 out <- bbt_run(
   fasta_filenames = get_babette_path("anthus_aco_sub.fas"),
-  mcmc = create_mcmc_nested_sampling(chain_length = 20000, store_every = 1000),
-  beast2_input_filename = beast2_input_file,
+  mcmc = create_mcmc_nested_sampling(
+    chain_length = 10000, 
+    store_every = 1000,
+    particle_count = 2,
+    sub_chain_length = 500
+  ),
   beast2_path = beastier::get_default_beast2_bin_path()
 )
 ```
