### 2018-09-13 b837392
Put BEAST2 output in 'bbt_run()$output'
