+++ tests/testthat/test-bbt_run.R
@@ -14,6 +14,7 @@ test_that("use, one alignment", {
   testthat::expect_true("estimates" %in% names(out))
   testthat::expect_true("anthus_aco_trees" %in% names(out))
   testthat::expect_true("operators" %in% names(out))
+  testthat::expect_true("output" %in% names(out))
   testthat::expect_equal(class(out$anthus_aco_trees[[1]]), "phylo")
   testthat::expect_equal(length(out$anthus_aco_trees), 2)
 
@@ -96,8 +97,14 @@ test_that("use, one alignment, same RNG should give same results", {
     mcmc = create_mcmc(chain_length = 1000, store_every = 1000),
     rng_seed = rng_seed
   )
+  # The screen output will be different here:
+  # [35] "Writing state to file /tmp/RtmpRyToHX/beast2_39f46a37b262.xml.state"
+  # [38] "File: beast2_39f41cfb21ef.xml seed: 42 threads: 1"
+  # [101] "Total calculation time: 0.197 seconds"
+  out_2$output[35] <- out_1$output[35]
+  out_2$output[38] <- out_1$output[38]
+  out_2$output[101] <- out_1$output[101]
   testthat::expect_identical(out_1, out_2)
-
 })
 
 test_that("use, two alignments, estimated crown ages", {
