+++ vignettes/nested_sampling.R
@@ -22,21 +22,23 @@ if (!mauricer::mrc_is_installed("NS"))
 }
 
 ## ------------------------------------------------------------------------
-beast2_input_file <- tempfile(fileext = ".xml")
-beautier::create_beast2_input_file(
-  input_filenames = get_babette_path("anthus_aco_sub.fas"),
-  output_filename = beast2_input_file,
-  mcmc = create_mcmc_nested_sampling(chain_length = 10000)
-)
-
-## ------------------------------------------------------------------------
 out <- bbt_run(
   fasta_filenames = get_babette_path("anthus_aco_sub.fas"),
-  mcmc = create_mcmc_nested_sampling(chain_length = 20000, store_every = 1000),
-  beast2_input_filename = beast2_input_file,
-  beast2_path = beastier::get_default_beast2_bin_path()
+  mcmc = create_mcmc_nested_sampling(
+    chain_length = 100000, 
+    store_every = -1,
+    particle_count = 1,
+    sub_chain_length = 500
+  ),
+  beast2_path = beastier::get_default_beast2_bin_path(),
+  beast2_input_filename = "/home/richel/out.xml",
+  cleanup = FALSE,
+  verbose = TRUE
 )
 
 ## ------------------------------------------------------------------------
 knitr::kable(out$estimates)
+
+## ------------------------------------------------------------------------
+print(out$output)
 
