+++ vignettes/nested_sampling.Rmd
@@ -49,12 +49,15 @@ if (!mauricer::mrc_is_installed("NS"))
 out <- bbt_run(
   fasta_filenames = get_babette_path("anthus_aco_sub.fas"),
   mcmc = create_mcmc_nested_sampling(
-    chain_length = 10000, 
-    store_every = 1000,
-    particle_count = 2,
+    chain_length = 100000, 
+    store_every = -1,
+    particle_count = 1,
     sub_chain_length = 500
   ),
-  beast2_path = beastier::get_default_beast2_bin_path()
+  beast2_path = beastier::get_default_beast2_bin_path(),
+  beast2_input_filename = "/home/richel/out.xml",
+  cleanup = FALSE,
+  verbose = TRUE
 )
 ```
 
@@ -63,6 +66,13 @@ Show the results, where `NSlikelihood` is the Nested-Sampling likelihood:
 ```{r}
 knitr::kable(out$estimates)
 ```
+
+Most is written to screen, however:
+
+```{r}
+print(out$output)
+```
+
 
 ## References
 
