### 2018-09-13 9a61c12
Test nested sampling
