+++ vignettes/nested_sampling.Rmd
@@ -54,23 +54,20 @@ out <- bbt_run(
     particle_count = 1,
     sub_chain_length = 500
   ),
-  beast2_path = beastier::get_default_beast2_bin_path(),
-  beast2_input_filename = "/home/richel/out.xml",
-  cleanup = FALSE,
-  verbose = TRUE
+  beast2_path = beastier::get_default_beast2_bin_path()
 )
 ```
 
 Show the results, where `NSlikelihood` is the Nested-Sampling likelihood:
 
 ```{r}
-knitr::kable(out$estimates)
+knitr::kable(tail(out$estimates))
 ```
 
-Most is written to screen, however:
+Marginal likelihood:
 
 ```{r}
-print(out$output)
+print(out$ns)
 ```
 
 
