+++ man/bbt_create_ns.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/bbt_create_ns.R
+\name{bbt_create_ns}
+\alias{bbt_create_ns}
+\title{Create the structure from a Nested Sampling run}
+\usage{
+bbt_create_ns(output)
+}
+\arguments{
+\item{output}{screen output}
+}
+\description{
+Create the structure from a Nested Sampling run
+}
+\author{
+Richel J.C. Bilderbeek
+}
