### 2018-09-17 d6dade5
@lintr-bot
