### 2018-09-19 4f73153
Improve docs, thanks Paula Quiroga, fix #25
