+++ .travis.yml
@@ -32,7 +32,7 @@ before_script:
 after_success:
   - Rscript -e 'lintr::lint_package()'
   - Rscript -e 'covr::codecov()'
-  - travis_wait 20 Rscript -e 'goodpractice::gp()'
+  - travis_wait 30 Rscript -e 'goodpractice::gp()'
   - proselint *.md | egrep -v "typography\.symbols\.curly_quotes"
   # temporarily disabled due to CERT_UNTRUSTED
   #- ./scripts/check_dead_links
