### 2018-09-19 25cad6d
Update vignettes
