+++ vignettes/nested_sampling.Rmd
@@ -17,7 +17,8 @@ knitr::opts_chunk$set(
 ```
 
 This vignette demonstrates how to use the Nested Sampling approach
-to obtain the marginal likelihood, as described in [1]
+to obtain the marginal likelihood and a Bayes factor, 
+as described in [1]
 
 ## Setup
 
@@ -28,11 +29,13 @@ library(babette)
 Check if BEAST2 is installed. If not, install it
 
 ```{r}
-if (!file.exists(get_default_beast2_path()))
+
+if (!beastier::is_beast2_installed())
 {
   print("Installing BEAST2 to default location")
-  install_beast2()
+  beastier::install_beast2()
 }
+testthat::expect_true(beastier::is_beast2_installed())
 ```
 
 ```{r}
@@ -41,28 +44,116 @@ if (!mauricer::mrc_is_installed("NS"))
   print("Installing BEAST2 package 'NS'")
   mauricer::mrc_install("NS")
 }
+testthat::expect_true(mauricer::mrc_is_installed("NS"))
+```
+
+Here we setup how to interpret the Bayes factor:
+```{r}
+interpret_bayes_factor <- function(bayes_factor) {
+  if (bayes_factor < 10^-2.0) {
+    "decisive for GTR"
+  } else if (bayes_factor < 10^-1.5) {
+    "very strong for GTR"
+  } else if (bayes_factor < 10^-1.0) {
+    "strong for GTR"
+  } else if (bayes_factor < 10^-0.5) {
+    "substantial for GTR"
+  } else if (bayes_factor < 10^0.0) {
+    "barely worth mentioning for GTR"
+  } else if (bayes_factor < 10^0.5) {
+    "barely worth mentioning for JC69"
+  } else if (bayes_factor < 10^1.0) {
+    "substantial for JC69"
+  } else if (bayes_factor < 10^1.5) {
+    "strong for JC69"
+  } else if (bayes_factor < 10^2.0) {
+    "very strong for JC69"
+  } else {
+    "decisive for JC69"
+  }
+}
+testit::assert(interpret_bayes_factor(1 / 123.0) == "decisive for GTR")
+testit::assert(interpret_bayes_factor(1 / 85.0) == "very strong for GTR")
+testit::assert(interpret_bayes_factor(1 / 12.5) == "strong for GTR")
+testit::assert(interpret_bayes_factor(1 / 8.5) == "substantial for GTR")
+testit::assert(interpret_bayes_factor(1 / 1.5) == "barely worth mentioning for GTR")
+testit::assert(interpret_bayes_factor(0.99) == "barely worth mentioning for GTR")
+testit::assert(interpret_bayes_factor(1.01) == "barely worth mentioning for JC69")
+testit::assert(interpret_bayes_factor(1.5) == "barely worth mentioning for JC69")
+testit::assert(interpret_bayes_factor(8.5) == "substantial for JC69")
+testit::assert(interpret_bayes_factor(12.5) == "strong for JC69")
+testit::assert(interpret_bayes_factor(85.0) == "very strong for JC69")
+testit::assert(interpret_bayes_factor(123.0) == "decisive for JC69")
+```
+
+## Experiment
+
+In this experiment, we will use the same DNA alignment to see which
+DNA nucleotide substitution model is the better fit.
+
+Load the DNA alignment, a subset of taxa from [2]:
+
+
+```{r}
+fasta_filename <- get_babette_path("anthus_aco_sub.fas")
+image(ape::read.FASTA(fasta_filename))
 ```
 
 ## Do the run
 
+In this vignette, the MCMC run is set up to be short:
+
 ```{r}
-out <- bbt_run(
[…]
