### 2018-09-22 8de8b1d
More tests on Travis, setup of https://github.com/richelbilderbeek/beautier/issues/46
