+++ tests/testthat/test-bbt_run.R
@@ -1,6 +1,7 @@
 context("bbt_run")
 
 test_that("use, one alignment", {
+  testit::assert(beastier::is_beast2_installed())
 
   out <- NA
 
@@ -40,6 +41,7 @@ test_that("use, one alignment", {
 })
 
 test_that("use, one alignment, verbose, cleanup", {
+  testit::assert(beastier::is_beast2_installed())
 
   testthat::expect_output(
     bbt_run(
@@ -51,6 +53,7 @@ test_that("use, one alignment, verbose, cleanup", {
 })
 
 test_that("use, one alignment, verbose, no cleanup", {
+  testit::assert(beastier::is_beast2_installed())
 
   fasta_filenames <- get_babette_path("anthus_aco.fas")
   beast2_input_filename <- tempfile(fileext = ".xml")
@@ -84,6 +87,7 @@ test_that("use, one alignment, verbose, no cleanup", {
 })
 
 test_that("use, one alignment, same RNG should give same results", {
+  testit::assert(beastier::is_beast2_installed())
 
   rng_seed <- 42
   out_1 <- bbt_run(
@@ -107,6 +111,7 @@ test_that("use, one alignment, same RNG should give same results", {
 })
 
 test_that("use, two alignments, estimated crown ages", {
+  testit::assert(beastier::is_beast2_installed())
 
   out <- NA
 
@@ -148,11 +153,7 @@ test_that("use, two alignments, estimated crown ages", {
   testthat::expect_true("rejectFC" %in% names(out$operators))
   testthat::expect_true("rejectIv" %in% names(out$operators))
   testthat::expect_true("rejectOp" %in% names(out$operators))
-
 })
-
-
-
 
 test_that("abuse", {
 
@@ -171,12 +172,13 @@ test_that("abuse", {
 ################################################################################
 
 test_that("Run all defaults", {
-
-  created <- beautier::create_beast2_input(
-    input_filenames = beautier::get_fasta_filename()
+  if (!beastier:::is_on_travis()) return()
+  expect_silent(
+    bbt_run(
+      fasta_filenames = get_beautier_path("anthus_aco_sub.fas"),
+      mcmc = beautier::create_mcmc(chain_length = 2000)
+    )
   )
-
-  testthat::expect_true(are_beast2_input_lines(created))
 })
 
 ################################################################################
@@ -188,28 +190,60 @@ test_that("Run all defaults", {
 ################################################################################
 
 test_that("Run GTR", {
-
-  created <- beautier::create_beast2_input(
-    input_filenames = beautier::get_fasta_filename(),
-    site_models = create_gtr_site_model()
+  if (!beastier:::is_on_travis()) return()
+  expect_silent(
+    bbt_run(
+      fasta_filenames = get_beautier_path("anthus_aco_sub.fas"),
+      site_models = create_gtr_site_model(),
+      mcmc = beautier::create_mcmc(chain_length = 2000)
+    )
   )
-
-  testthat::expect_true(are_beast2_input_lines(created))
 })
 
 ################################################################################
 # Site model: HKY
 ################################################################################
 
+test_that("Run HKY", {
+  if (!beastier:::is_on_travis()) return()
+  expect_silent(
+    bbt_run(
+      fasta_filenames = get_beautier_path("anthus_aco_sub.fas"),
+      site_models = create_hky_site_model(),
+      mcmc = beautier::create_mcmc(chain_length = 2000)
+    )
+  )
+})
+
 ################################################################################
 # Site model: JC69
 ################################################################################
 
+test_that("Run JC69", {
+  if (!beastier:::is_on_travis()) return()
+  expect_silent(
+    bbt_run(
[…]
