+++ vignettes/nested_sampling.R
@@ -8,11 +8,13 @@ knitr::opts_chunk$set(
 library(babette)
 
 ## ------------------------------------------------------------------------
-if (!file.exists(get_default_beast2_path()))
+
+if (!beastier::is_beast2_installed())
 {
   print("Installing BEAST2 to default location")
-  install_beast2()
+  beastier::install_beast2()
 }
+testthat::expect_true(beastier::is_beast2_installed())
 
 ## ------------------------------------------------------------------------
 if (!mauricer::mrc_is_installed("NS"))
@@ -20,19 +22,80 @@ if (!mauricer::mrc_is_installed("NS"))
   print("Installing BEAST2 package 'NS'")
   mauricer::mrc_install("NS")
 }
+testthat::expect_true(mauricer::mrc_is_installed("NS"))
 
 ## ------------------------------------------------------------------------
-out <- bbt_run(
-  fasta_filenames = get_babette_path("anthus_aco_sub.fas"),
-  mcmc = create_mcmc_nested_sampling(
-    chain_length = 100000, 
-    store_every = -1,
-    particle_count = 1,
-    sub_chain_length = 500
-  ),
+interpret_bayes_factor <- function(bayes_factor) {
+  if (bayes_factor < 10^-2.0) {
+    "decisive for GTR"
+  } else if (bayes_factor < 10^-1.5) {
+    "very strong for GTR"
+  } else if (bayes_factor < 10^-1.0) {
+    "strong for GTR"
+  } else if (bayes_factor < 10^-0.5) {
+    "substantial for GTR"
+  } else if (bayes_factor < 10^0.0) {
+    "barely worth mentioning for GTR"
+  } else if (bayes_factor < 10^0.5) {
+    "barely worth mentioning for JC69"
+  } else if (bayes_factor < 10^1.0) {
+    "substantial for JC69"
+  } else if (bayes_factor < 10^1.5) {
+    "strong for JC69"
+  } else if (bayes_factor < 10^2.0) {
+    "very strong for JC69"
+  } else {
+    "decisive for JC69"
+  }
+}
+testit::assert(interpret_bayes_factor(1 / 123.0) == "decisive for GTR")
+testit::assert(interpret_bayes_factor(1 / 85.0) == "very strong for GTR")
+testit::assert(interpret_bayes_factor(1 / 12.5) == "strong for GTR")
+testit::assert(interpret_bayes_factor(1 / 8.5) == "substantial for GTR")
+testit::assert(interpret_bayes_factor(1 / 1.5) == "barely worth mentioning for GTR")
+testit::assert(interpret_bayes_factor(0.99) == "barely worth mentioning for GTR")
+testit::assert(interpret_bayes_factor(1.01) == "barely worth mentioning for JC69")
+testit::assert(interpret_bayes_factor(1.5) == "barely worth mentioning for JC69")
+testit::assert(interpret_bayes_factor(8.5) == "substantial for JC69")
+testit::assert(interpret_bayes_factor(12.5) == "strong for JC69")
+testit::assert(interpret_bayes_factor(85.0) == "very strong for JC69")
+testit::assert(interpret_bayes_factor(123.0) == "decisive for JC69")
+
+## ------------------------------------------------------------------------
+fasta_filename <- get_babette_path("anthus_aco_sub.fas")
+image(ape::read.FASTA(fasta_filename))
+
+## ------------------------------------------------------------------------
+mcmc <- beautier::create_mcmc_nested_sampling(
+  chain_length = 10000, 
+  store_every = -1,
+  particle_count = 1,
+  sub_chain_length = 500
+)
+
[…]
