### 2018-10-11 3e5f440
Add 'bbt_create_test_ns_output'
