+++ tests/testthat/test-bbt_create_test_ns_output.R
@@ -1,0 +1,7 @@
+context("bbt_create_test_ns_output")
+
+test_that("use", {
+  expect_true(
+    length(babette:::bbt_create_test_ns_output()) > 100
+  )
+})
