### 2018-10-20 e73a76a
AppVeyor installs mauricer
