### 2018-10-20 4078a62
No Nested Sampling run under Windows
