### 2018-10-20 0919bd3
Tests pass on Linux
