### 2018-10-20 be022e0
Fix spelling errors
