+++ R/bbt_create_ns.R
@@ -2,9 +2,9 @@
 #' @param output screen output
 #' @return a list with the following elements:
 #'   \itemize{
-#'     \item marg_log_lik the marginal log likelihood estimate
-#'     \item marg_log_lik_sd the standard deviation around the estimate
-#'     \item ess the effective sample size
+#'     \item \code{marg_log_lik} the marginal log likelihood estimate
+#'     \item \code{marg_log_lik_sd} the standard deviation around the estimate
+#'     \item \code{ess} the effective sample size
 #'       in estimating the marginal log likelihood
 #'   }
 #' @author Richel J.C. Bilderbeek
