+++ man/bbt_create_ns.Rd
@@ -12,9 +12,9 @@ bbt_create_ns(output)
 \value{
 a list with the following elements:
   \itemize{
-    \item marg_log_lik the marginal log likelihood estimate
-    \item marg_log_lik_sd the standard deviation around the estimate
-    \item ess the effective sample size
+    \item \code{marg_log_lik} the marginal log likelihood estimate
+    \item \code{marg_log_lik_sd} the standard deviation around the estimate
+    \item \code{ess} the effective sample size
       in estimating the marginal log likelihood
   }
 }
