### 2018-10-20 378b475
v1.2.5 candidate [skip ci]
