+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: babette
 Title: Control BEAST2 from R
-Version: 1.2.4
+Version: 1.2.5
 Authors@R: c(
   person("Richel", "Bilderbeek", email = "richel@richelbilderbeek.nl", role = c("aut", "cre"), comment = c(ORCID = "0000-0003-1107-7049")),
   person("Joelle", "Barido-Sottani", role = "rev", comment = "Joelle reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/209"),
