### 2018-10-25 44672ae
Use simplified interface without hyper parameters
