+++ NEWS.md
@@ -2,6 +2,36 @@
 
 Newest versions at top.
 
+## babette 1.2.5 (2018-10-25)
+
+### NEW FEATURES
+
+  * None
+
+### MINOR IMPROVEMENTS
+
+  * Simplified interface for parameters:
+
+```
+# Old
+distr <- create_distr_poisson(id = 1, lambda = create_lambda_param(value = 1.2))
+# Added
+distr <- create_distr_poisson(id = 1, lambda = 1.2)
+```
+
+### BUG FIXES
+
+  * None
+
+### DEPRECATED AND DEFUNCT
+
+  * Cannot use hyper parameters
+
+```
+# Deprecated: create a hyper parameter
+create_lambda_param(value = 1.2, estimate = TRUE)
+```
+
 ## babette 1.2.4 (2018-05-17)
 
 ### NEW FEATURES
