### 2018-10-26 #209 @richelbilderbeek
All feedback has been addressed, last (and biggest) part was about [beautier](https://github.com/richelbilderbeek/beautier/projects/1). Rounding things off now, then will let know.
