+++ doc/faq.md
@@ -66,7 +66,7 @@ gives prettier results.
 ### How do I reference to this work?
 
 ```
-Bilderbeek, Richel JC, and Rampal S. Etienne. "babette: BEAU ti 2, BEAST 2 and Tracer for R." Methods in Ecology and Evolution (2018).
+Bilderbeek, Richel JC, and Rampal S. Etienne. "babette: BEAUti 2, BEAST 2 and Tracer for R." Methods in Ecology and Evolution (2018).
 ```
 
 or
