### 2018-10-26 richelbilderbeek/babette#28 @richelbilderbeek
Feature request: multiple alignments
Moved from [there](https://github.com/richelbilderbeek/beautier/issues/69):

[Feedback dwinter](https://github.com/ropensci/onboarding/issues/209#issuecomment-399874987):

> As far as I can tell it is not possible to estimate one tree from multiple alignments. (i.e. one tree prior, multiple alignments possibly each with its own site model). I suspect this is a common use-case (either specified this way or through a concatenated alignment with partitions). If I have this wrong then a partitioned dataset should be added to the examples vignette. If this is not currently possible it should be a priority for future development
