### 2018-10-26 #209 @richelbilderbeek
@noamross: I have processed all feedback! Blimey, the feedback from @dwinter and @bjoelle heavily improved the packages. Sure, I've put both of them in the DESCRIPTION file of all packages.

So, where do we go from here? :+1:
