### 2018-10-29 f393e1c
Only use Authors field
