### 2018-10-29 20d5c5c
Use simpler interface
