### 2018-10-29 118e25c
Update doc and vignettes
