+++ vignettes/tutorial.R
@@ -14,44 +14,17 @@ testit::assert(file.exists(fasta_filename))
 ## ------------------------------------------------------------------------
 mcmc <- create_mcmc(chain_length = 2000, store_every = 1000)
 
-## ----cache=TRUE----------------------------------------------------------
-out <- bbt_run(
- fasta_filenames = fasta_filename,
- mcmc = mcmc
-)
-
 ## ------------------------------------------------------------------------
 site_model <- create_site_model_jc69()
 site_model <- create_jc69_site_model()
-
-## ----cache=TRUE----------------------------------------------------------
-out <- bbt_run(
-  fasta_filenames = fasta_filename,
-  site_models = site_model,
-  mcmc = mcmc
-)
 
 ## ------------------------------------------------------------------------
 clock_model <- create_clock_model_strict()
 clock_model <- create_strict_clock_model()
 
-## ----cache=TRUE----------------------------------------------------------
-out <- bbt_run(
-  fasta_filenames = fasta_filename,
-  clock_models = clock_model,
-  mcmc = mcmc
-)
-
 ## ------------------------------------------------------------------------
 tree_prior <- create_tree_prior_yule()
 tree_prior <- create_yule_tree_prior()
-
-## ----cache=TRUE----------------------------------------------------------
-out <- bbt_run(
-  fasta_filenames = fasta_filename,
-  tree_priors = tree_prior,
-  mcmc = mcmc
-)
 
 ## ------------------------------------------------------------------------
 mrca_prior <- create_mrca_prior(
@@ -62,8 +35,8 @@ mrca_prior <- create_mrca_prior(
 
 ## ------------------------------------------------------------------------
 mrca_distr <- create_normal_distr(
-  mean = create_mean_param(value = 15.0), 
-  sigma = create_sigma_param(value = 1.0)
+  mean = 15.0, 
+  sigma = 1.0
 )
 
 ## ------------------------------------------------------------------------
@@ -74,21 +47,6 @@ mrca_prior <- create_mrca_prior(
 )
 
 ## ----cache=TRUE----------------------------------------------------------
-out <- bbt_run(
-  fasta_filenames = fasta_filename,
-  mrca_priors = mrca_prior,
-  mcmc = mcmc
-)
-
-## ----cache=TRUE----------------------------------------------------------
-# Prefer to use the 'create_mrca_prior' version below
-out <- bbt_run(
-  fasta_filenames = fasta_filename,
-  posterior_crown_age = 15,
-  mcmc = mcmc
-)
-
-## ----cache=TRUE----------------------------------------------------------
 # Prefer this over using 'posterior_crown_age'
 out <- bbt_run(
   fasta_filenames = fasta_filename,
@@ -96,8 +54,8 @@ out <- bbt_run(
     alignment_id = get_alignment_id(fasta_filename = fasta_filename), 
     taxa_names = get_taxa_names(filename = fasta_filename),
     mrca_distr = create_normal_distr(
-      mean = create_mean_param(value = 15.0), 
-      sigma = create_sigma_param(value = 0.0001)
+      mean = 15.0, 
+      sigma = 0.0001
     )
   ),
   mcmc = mcmc
@@ -128,13 +86,6 @@ if (1 == 2) {
   file.remove(all_files)
 }
 
-## ----cache=TRUE----------------------------------------------------------
-out <- bbt_run(
-  fasta_filenames = fasta_filename,
-  mcmc = mcmc,
-  rng_seed = 314
-)
-
 ## ------------------------------------------------------------------------
 beast2_path <- beastier::get_default_beast2_path()
 print(beast2_path)
[…]
