+++ vignettes/examples.R
@@ -53,8 +53,8 @@ posterior <- bbt_run(
     alignment_id = get_alignment_id("my_fasta.fas"),
     is_monophyletic = TRUE,
     mrca_distr = create_normal_distr(
-      mean = create_mean_param(value = 15.0, estimate = FALSE),
-      sigma = create_sigma_param(value = 0.025, estimate = FALSE)
+      mean = 15.0,
+      sigma = 0.025
     )
   )
 )
@@ -113,8 +113,8 @@ posterior <- bbt_run(
   "my_alignment.fas",
   tree_priors = create_yule_tree_prior(
     birth_rate_distr = create_normal_distr(
-      mean = create_mean_param(value = 1.0),
-      sigma = create_sigma_param(value = 0.1)
+      mean = 1.0,
+      sigma = 0.1
     )
   ),
   mcmc = mcmc
@@ -147,7 +147,7 @@ plot_densitree(posterior$my_alignment_trees, width = 2)
 posterior <- bbt_run(
   "my_alignment.fas",
   clock_models = create_strict_clock_model(
-    clock_rate_param = create_clock_rate_param(value = 0.5)
+    clock_rate_param = 0.5
   ), 
   mcmc = mcmc
 )
