+++ vignettes/tutorial.Rmd
@@ -221,8 +221,8 @@ use `create_normal_distr`:
 
 ```{r}
 mrca_distr <- create_normal_distr(
-  mean = create_mean_param(value = 15.0), 
-  sigma = create_sigma_param(value = 1.0)
+  mean = 15.0, 
+  sigma = 1.0
 )
 ```
 
@@ -271,8 +271,8 @@ out <- bbt_run(
     alignment_id = get_alignment_id(fasta_filename = fasta_filename), 
     taxa_names = get_taxa_names(filename = fasta_filename),
     mrca_distr = create_normal_distr(
-      mean = create_mean_param(value = 15.0), 
-      sigma = create_sigma_param(value = 0.0001)
+      mean = 15.0, 
+      sigma = 0.0001
     )
   ),
   mcmc = mcmc
