+++ .travis.yml
@@ -33,6 +33,7 @@ after_success:
   - Rscript -e 'lintr::lint_package()'
   - Rscript -e 'covr::codecov()'
   - proselint *.md | egrep -v "typography\.symbols\.curly_quotes"
+  - cd scripts; ./build_cran
   #- travis_wait 30 Rscript -e 'goodpractice::gp()'
   # temporarily disabled due to CERT_UNTRUSTED
   #- ./scripts/check_dead_links
