### 2018-10-30 8efcf52
Improve doc [skip ci]
