### 2018-10-30 e419dd9
Use R 3.3 for MacOS
