### 2018-10-30 5012c78
Use default R version
