+++ .travis.yml
@@ -1,9 +1,6 @@
 language: r
 cache: packages
 
-r:
-  - 3.3
-
 os: 
   - osx
   - linux
