### 2018-10-30 25f0fda
Workaround to get OSX build working, from https://github.com/travis-ci/travis-ci/issues/10295
