### 2018-10-31 22bbd19
Test for local temporaries created
