+++ .travis.yml
@@ -41,6 +41,10 @@ after_success:
   - Rscript -e 'lintr::lint_package()'
   - Rscript -e 'covr::codecov()'
   - proselint *.md | egrep -v "typography\.symbols\.curly_quotes"
+  - # No local files created
+  - egrep -oR "\"~" --include={*.R,*.Rmd}
+  - if [[ $(egrep -R "\"~" --include={*.R,*.Rmd}) ]]; then echo "Local file created"; exit 1; fi
+  - # Do a CRAN build
   - cd scripts; ./build_cran
   #- travis_wait 30 Rscript -e 'goodpractice::gp()'
   # temporarily disabled due to CERT_UNTRUSTED
