### 2018-11-01 06bc9d2
Only search for locally created temporaries on Linux
