+++ .travis.yml
@@ -49,8 +49,7 @@ after_success:
   - Rscript -e 'covr::codecov()'
   - proselint *.md | egrep -v "typography\.symbols\.curly_quotes"
   - # No local files created
-  - egrep -oR "\"~" --include={*.R,*.Rmd}
-  - if [[ $(egrep -R "\"~" --include={*.R,*.Rmd}) ]]; then echo "Local file created"; exit 1; fi
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then if [[ $(egrep -R "\"~" --include={*.R,*.Rmd}) ]]; then echo "Local file created"; exit 1; fi; fi
   - # Do a CRAN build
   - cd scripts; ./build_cran
   #- travis_wait 30 Rscript -e 'goodpractice::gp()'
