### 2018-11-01 d742b06
Show AppVeyor status of dependencies [skip ci]
