### 2018-11-01 fa690d9
Show AppVeyor status of dependencies [skip ci]
