### 2018-11-01 e4556ae
Ignore CRAN build folder
