+++ scripts/build/babette.Rcheck/00_pkg_src/babette/R/zzz.R
@@ -1,12 +1,0 @@
-.onLoad <- function(libname, pkgname){
-
-  suppressPackageStartupMessages(
-    lapply(
-      c("beautier", "beastier", "tracerer"),
-      library,
-      character.only = TRUE,
-      warn.conflicts = FALSE
-    )
-  )
-  invisible()
-}
