+++ scripts/build/babette.Rcheck/00_pkg_src/babette/R/bbt_run.R
@@ -1,242 +1,0 @@
-#' Do a full run: create a BEAST2 configuration file (like BEAUti 2),
-#' run BEAST2, parse results (like Tracer)
-#' @inheritParams bbt_default_params_doc
-#' @return a list with the following elements:\cr
-#' \itemize{
-#'   \item{
-#'     \code{estimates}: a data frame with BEAST2
-#'     parameter estimates
-#'   }
-#'   \item{
-#'     \code{[alignment_id]_trees}: a \code{multiPhylo}
-#'     containing the phylogenies
-#'     in the BEAST2 posterior. \code{[alignment_id]}
-#'     is the ID of the alignment. For example,
-#'     when running \code{bbt_run} with
-#'     \code{anthus_aco.fas}, this element will have
-#'     name \code{anthus_aco_trees}
-#'   }
-#'   \item{
-#'     \code{operators}: a data frame with the
-#'     BEAST2 MCMC operator acceptances
-#'   }
-#'   \item{
-#'     \code{output}: a numeric vector with the output
-#'     sent to standard output and error streams
-#'   }
-#'   \item{
-#'     \code{ns}: (optional) the results of the NS BEAST2 package,
-#'     will exist only when \code{create_mcmc_nested_sampling} was
-#'     used for \code{mcmc}
-#'   }
-#' }
-#' @author Richel J.C. Bilderbeek
-#' @usage
-#' bbt_run(
-#'   fasta_filenames,
-#'   site_models = beautier::create_jc69_site_models(
-#'     beautier::get_alignment_ids(fasta_filenames)
-#'   ),
-#'   clock_models = beautier::create_strict_clock_models(
-#'     beautier::get_alignment_ids(fasta_filenames)
-#'   ),
-#'   tree_priors = beautier::create_yule_tree_priors(
-#'     beautier::get_alignment_ids(fasta_filenames)
-#'   ),
-#'   mrca_priors = NA,
-#'   mcmc = beautier::create_mcmc(),
-#'   posterior_crown_age = NA,
-#'   beast2_input_filename = tempfile(pattern = "beast2_", fileext = ".xml"),
-#'   rng_seed = 1,
-#'   beast2_output_log_filename = tempfile(
-#'     pattern = "beast2_", fileext = "log"
-#'   ),
-#'   beast2_output_trees_filenames = tempfile(
-#'     pattern = paste0(
-#'       "beast2_",
-#'       beautier::get_alignment_ids(fasta_filenames), "_"
-#'     ),
-#'     fileext = ".trees"
-#'   ),
-#'   beast2_output_state_filename = tempfile(
-#'     pattern = "beast2_", fileext = ".xml.state"
-#'   ),
-#'   beast2_path = beastier::get_default_beast2_path(),
-#'   verbose = FALSE,
-#'   cleanup = TRUE
-#' )
-#' @examples
-#'  # One alignment
-#'  out <- bbt_run(
-#'    fasta_filenames = get_babette_path("anthus_aco.fas"),
-#'    mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
-#'  )
-#'
-#'  testit::assert("estimates" %in% names(out))
-#'  testit::assert("anthus_aco_trees" %in% names(out))
-#'  testit::assert("operators" %in% names(out))
-#'  testit::assert("output" %in% names(out))
-#'  testit::assert(class(out$anthus_aco_trees[[1]]) == "phylo")
-#'  testit::assert(length(out$anthus_aco_trees) == 2)
-#'
-#'  testit::assert("Sample" %in% names(out$estimates))
-#'  testit::assert("posterior" %in% names(out$estimates))
-#'  testit::assert("likelihood" %in% names(out$estimates))
-#'  testit::assert("prior" %in% names(out$estimates))
[…]
