+++ scripts/build/babette.Rcheck/00_pkg_src/babette/R/bbt_process_pkg_output.R
@@ -1,12 +1,0 @@
-#' Process the BEAST2 output dependent on BEAST2 package specifics
-#' @param out complete babette output
-#' @param mcmc an MCMC
-bbt_process_pkg_output <- function(
-  out,
-  mcmc
-) {
-  if (beautier:::is_mcmc_nested_sampling(mcmc)) {
-    out$ns <- bbt_create_ns(out$output) # nolint internal function
-  }
-  out
-}

[…]
