### 2018-11-01 7c9f86e
Build HTML vignette
