### 2018-11-01 5b171a2
No CRAN build on Travis
