### 2018-11-01 55b463b
RNG seed unreliable under macOS
