### 2018-11-01 1a485b7
Test less on macOS, as it is slower
