+++ .travis.yml
@@ -43,9 +43,9 @@ script:
   - travis_wait 30 R CMD check *tar.gz
 
 after_success:
-  - Rscript -e 'lintr::lint_package()'
-  - Rscript -e 'covr::codecov()'
-  - proselint *.md | egrep -v "typography\.symbols\.curly_quotes"
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then Rscript -e 'lintr::lint_package()'; fi
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then Rscript -e 'covr::codecov()'; fi
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then proselint *.md | egrep -v "typography\.symbols\.curly_quotes"; fi
   - # No local files created
   - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then if [[ $(egrep -R "\"~" --include={*.R,*.Rmd}) ]]; then echo "Local file created"; exit 1; fi; fi
   # Do a CRAN build. No time :-(
