### 2018-11-08 be83a84
Proper error message if using Nested Sampling and the BEAST2 jar file, fix #30
