+++ R/bbt_check_beast2_packages.R
@@ -13,4 +13,14 @@ bbt_check_beast2_packages <- function(
       "Tip: use 'mauricer::mrc_install(\"NS\")'"
     )
   }
+  if (beautier:::is_mcmc_nested_sampling(mcmc) &&
+      !beastier:::is_bin_path(beast2_path)
+  ) {
+    stop(
+      "When using nested sampling (that is, ",
+      "using 'create_mcmc_nested_sampling'), ",
+      "one must use the binary BEAST2 executable (that is, using ",
+      "'beast2_path = get_default_beast2_bin_path()')"
+    )
+  }
 }
