+++ tests/testthat/test-bbt_run.R
@@ -405,7 +405,7 @@ test_that("Run MRCA, no distr, subset of taxa", {
   testthat::expect_true(are_beast2_input_lines(lines))
 })
 
-test_that("RLN and non-monophyletic MRCA, Issue 29, #29", {
+test_that("RLN and non-monophyletic MRCA with distribution, Issue 29, #29", {
 
   skip("Invalid BEAST2 file when using RLN and MRCA prior, Issue 29, #29")
   # Thanks to Raphael Scherrer for sharing this bug
@@ -417,7 +417,7 @@ test_that("RLN and non-monophyletic MRCA, Issue 29, #29", {
       alignment_id = get_alignment_id(fasta_filename),
       taxa_names = get_taxa_names(fasta_filename),
       is_monophyletic = FALSE,
-      mrca_distr = create_one_div_x_distr()
+      mrca_distr = create_one_div_x_distr() # Use simpler distribution
     )
   )
   testthat::expect_true(
@@ -427,7 +427,7 @@ test_that("RLN and non-monophyletic MRCA, Issue 29, #29", {
   )
 })
 
-test_that("RLN and monophyletic MRCA, Issue 29, #29", {
+test_that("RLN and monophyletic MRCA with distribution, Issue 29, #29", {
 
   skip("Invalid BEAST2 file when using RLN and MRCA prior, Issue 29, #29")
 
@@ -440,7 +440,7 @@ test_that("RLN and monophyletic MRCA, Issue 29, #29", {
       alignment_id = get_alignment_id(fasta_filename),
       taxa_names = get_taxa_names(fasta_filename),
       is_monophyletic = TRUE,
-      mrca_distr = create_one_div_x_distr()
+      mrca_distr = create_one_div_x_distr() # Use simpler distributiob
     )
   )
   testthat::expect_true(
