### 2018-11-10 195127f
No support for multiple alignments anymore
