+++ vignettes/article.Rmd
@@ -52,14 +52,17 @@ out <- bbt_run(
 
 
 ```{r}
-out <- bbt_run(
-  c("anthus_aco.fas", "anthus_nd2.fas"),
-  site_models = list(
-    create_tn93_site_model(), 
-    create_gtr_site_model()
-  ),
-  mcmc = mcmc
-)
+# Currently, support for multiple alignments has been deprecated
+if (1 == 2) {
+  out <- bbt_run(
+    c("anthus_aco.fas", "anthus_nd2.fas"),
+    site_models = list(
+      create_tn93_site_model(), 
+      create_gtr_site_model()
+    ),
+    mcmc = mcmc
+  )
+}
 ```
 
 ```{r}
