+++ vignettes/examples.Rmd
@@ -260,58 +260,6 @@ plot_densitree(posterior$my_alignment_trees, width = 2)
 
 Thanks to Paul van Els and Yacine Ben Chehida for this use case.
 
-## Example #9: Two alignments
-
-![Example 9: Two alignments](anthus_2_4.png)
-
-```{r example_9, cache=TRUE}
-posterior <- bbt_run(
-  c("anthus_aco.fas", "anthus_nd2.fas"),
-  mcmc = mcmc
-)
-```
-
-```{r fig.width=7, fig.height=7}
-ggplot2::ggplot(
-  data = posterior$estimates,
-  ggplot2::aes(x = Sample)
-) + ggplot2::geom_line(ggplot2::aes(y = birthRate.aco), color = "red") +
-  ggplot2::geom_line(ggplot2::aes(y = birthRate.nd2), color = "green")
-plot_densitree(posterior$anthus_aco_trees, width = 2)
-plot_densitree(posterior$anthus_nd2_trees, width = 2)
-```
-
-Thanks to Paul van Els for this use case and supplying these FASTA files.
-
-## Example #10: Two alignments, different site models
-
-![Example 10: Two alignments, different site models](aco_hky_nd2_tn93.png)
-
-```{r example_10, cache=TRUE}
-posterior <- bbt_run(
-  c("anthus_aco.fas", "anthus_nd2.fas"),
-  site_models = list(
-    create_hky_site_model(), 
-    create_tn93_site_model()
-  ),
-  mcmc = mcmc
-)
-```
-
-```{r fig.width=7, fig.height=7}
-ggplot2::ggplot(
-  data = posterior$estimates,
-  ggplot2::aes(x = Sample)
-) + ggplot2::geom_line(ggplot2::aes(y = birthRate.aco), color = "red") +
-  ggplot2::geom_line(ggplot2::aes(y = birthRate.nd2), color = "green")
-plot_densitree(posterior$anthus_aco_trees, width = 2)
-plot_densitree(posterior$anthus_nd2_trees, width = 2)
-```
-
-Since `v1.12` this it is supported to have two alignments with different site models, clock models and tree priors.
-
-Thanks to Paul van Els for this use case.
-
 ```{r cleanup, include = FALSE}
 file.remove("test_output_0.fas")
 file.remove("my_fasta.fas")
