+++ tests/testthat/test-bbt_run.R
@@ -451,108 +451,8 @@ test_that("RLN and monophyletic MRCA with distribution, Issue 29, #29", {
 })
 
 ################################################################################
-# Initial phylogenies
+# nLTT
 ################################################################################
-
-test_that("JC69 JC69 strict strict coalescent_exp_population", {
-
-  input_fasta_filename_1 <- get_babette_path("anthus_aco.fas")
-  input_fasta_filename_2 <- get_babette_path("anthus_nd2.fas")
-  input_filenames <- c(input_fasta_filename_1, input_fasta_filename_2)
-  site_model_1 <- create_jc69_site_model()
-  site_model_2 <- create_jc69_site_model()
-  clock_model_1 <- create_strict_clock_model()
-  clock_model_2 <- create_strict_clock_model()
-  tree_prior <- create_cep_tree_prior()
-  lines <- create_beast2_input(
-    input_filenames = input_filenames,
-    site_models = list(site_model_1, site_model_2),
-    clock_models = list(clock_model_1, clock_model_2),
-    tree_priors = list(tree_prior, tree_prior)
-  )
-  testthat::expect_true(are_beast2_input_lines(lines))
-})
-
-test_that("TN93 TN93 strict strict yule", {
-
-  input_fasta_filename_1 <- get_babette_path("anthus_aco.fas")
-  input_fasta_filename_2 <- get_babette_path("anthus_nd2.fas")
-  input_filenames <- c(input_fasta_filename_1, input_fasta_filename_2)
-  site_model_1 <- create_tn93_site_model()
-  site_model_2 <- create_tn93_site_model()
-  clock_model_1 <- create_strict_clock_model()
-  clock_model_2 <- create_strict_clock_model()
-  tree_prior <- create_yule_tree_prior()
-  lines <- create_beast2_input(
-    input_filenames = input_filenames,
-    site_models = list(site_model_1, site_model_2),
-    clock_models = list(clock_model_1, clock_model_2),
-    tree_priors = list(tree_prior, tree_prior)
-  )
-  testthat::expect_true(are_beast2_input_lines(lines))
-})
-
-
-
-test_that("GTR GTR strict strict yule", {
-
-  input_fasta_filename_1 <- get_babette_path("anthus_aco.fas")
-  input_fasta_filename_2 <- get_babette_path("anthus_nd2.fas")
-  input_filenames <- c(input_fasta_filename_1, input_fasta_filename_2)
-  site_model_1 <- create_gtr_site_model()
-  site_model_2 <- create_gtr_site_model()
-  clock_model_1 <- create_strict_clock_model()
-  clock_model_2 <- create_strict_clock_model()
-  tree_prior <- create_yule_tree_prior()
-  lines <- create_beast2_input(
-    input_filenames = input_filenames,
-    site_models = list(site_model_1, site_model_2),
-    clock_models = list(clock_model_1, clock_model_2),
-    tree_priors = list(tree_prior, tree_prior)
-  )
-  testthat::expect_true(are_beast2_input_lines(lines))
-})
-
-
-test_that("GTR TN93 strict strict yule", {
-
-  input_fasta_filename_1 <- get_babette_path("anthus_aco.fas")
-  input_fasta_filename_2 <- get_babette_path("anthus_nd2.fas")
-  input_filenames <- c(input_fasta_filename_1, input_fasta_filename_2)
-  site_model_1 <- create_gtr_site_model()
-  site_model_2 <- create_tn93_site_model()
-  clock_model_1 <- create_strict_clock_model()
-  clock_model_2 <- create_strict_clock_model()
-  tree_prior <- create_yule_tree_prior()
-  lines <- create_beast2_input(
[…]
