### 2018-11-10 5d26e9b
Test for #29 being fixed
