+++ tests/testthat/test-bbt_run.R
@@ -115,51 +115,6 @@ test_that("use, one alignment, same RNG should give same results", {
   testthat::expect_identical(out_1, out_2)
 })
 
-test_that("use, two alignments, estimated crown ages", {
-  testit::assert(beastier::is_beast2_installed())
-
-  out <- NA
-
-  testthat::expect_silent(
-    out <- bbt_run(
-      fasta_filenames = get_babette_paths(
-        c("anthus_aco.fas", "anthus_nd2.fas")
-      ),
-      mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
-    )
-  )
-  testthat::expect_true("estimates" %in% names(out))
-  testthat::expect_true("anthus_aco_trees" %in% names(out))
-  testthat::expect_true("anthus_nd2_trees" %in% names(out))
-  testthat::expect_true("operators" %in% names(out))
-  testthat::expect_equal(class(out$anthus_aco_trees[[1]]), "phylo")
-  testthat::expect_equal(class(out$anthus_nd2_trees[[1]]), "phylo")
-  testthat::expect_equal(length(out$anthus_aco_trees), 2)
-  testthat::expect_equal(length(out$anthus_nd2_trees), 2)
-
-  testthat::expect_true("Sample" %in% names(out$estimates))
-  testthat::expect_true("posterior" %in% names(out$estimates))
-  testthat::expect_true("likelihood" %in% names(out$estimates))
-  testthat::expect_true("prior" %in% names(out$estimates))
-  testthat::expect_true("treeLikelihood.aco" %in% names(out$estimates))
-  testthat::expect_true("treeLikelihood.nd2" %in% names(out$estimates))
-  testthat::expect_true("TreeHeight.aco" %in% names(out$estimates))
-  testthat::expect_true("TreeHeight.nd2" %in% names(out$estimates))
-  testthat::expect_true("YuleModel.aco" %in% names(out$estimates))
-  testthat::expect_true("YuleModel.nd2" %in% names(out$estimates))
-  testthat::expect_true("birthRate.aco" %in% names(out$estimates))
-  testthat::expect_true("birthRate.nd2" %in% names(out$estimates))
-
-  testthat::expect_true("operator" %in% names(out$operators))
-  testthat::expect_true("p" %in% names(out$operators))
-  testthat::expect_true("accept" %in% names(out$operators))
-  testthat::expect_true("reject" %in% names(out$operators))
-  testthat::expect_true("acceptFC" %in% names(out$operators))
-  testthat::expect_true("rejectFC" %in% names(out$operators))
-  testthat::expect_true("rejectIv" %in% names(out$operators))
-  testthat::expect_true("rejectOp" %in% names(out$operators))
-})
-
 test_that("abuse", {
 
   testthat::expect_error(
