### 2018-11-13 e438ddb
bbt_run uses a list of one element
