+++ R/bbt_run.R
@@ -34,14 +34,20 @@
 #' @usage
 #' bbt_run(
 #'   fasta_filenames,
-#'   site_models = beautier::create_jc69_site_models(
-#'     beautier::get_alignment_ids(fasta_filenames)
+#'   site_models = list(
+#'     beautier::create_jc69_site_model(
+#'       beautier::get_alignment_id(fasta_filenames)
+#'     )
 #'   ),
-#'   clock_models = beautier::create_strict_clock_models(
-#'     beautier::get_alignment_ids(fasta_filenames)
+#'   clock_models = list(
+#'     beautier::create_strict_clock_model(
+#'       beautier::get_alignment_ids(fasta_filenames)
+#'     )
 #'   ),
-#'   tree_priors = beautier::create_yule_tree_priors(
-#'     beautier::get_alignment_ids(fasta_filenames)
+#'   tree_priors = list(
+#'     beautier::create_yule_tree_prior(
+#'       beautier::get_alignment_ids(fasta_filenames)
+#'     )
 #'   ),
 #'   mrca_priors = NA,
 #'   mcmc = beautier::create_mcmc(),
@@ -66,7 +72,6 @@
 #'   cleanup = TRUE
 #' )
 #' @examples
-#'  # One alignment
 #'  out <- bbt_run(
 #'    fasta_filenames = get_babette_path("anthus_aco.fas"),
 #'    mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
@@ -102,12 +107,21 @@
 #' @export
 bbt_run <- function(
   fasta_filenames,
-  site_models = beautier::create_jc69_site_models(
-    beautier::get_alignment_ids(fasta_filenames)),
-  clock_models = beautier::create_strict_clock_models(
-    beautier::get_alignment_ids(fasta_filenames)),
-  tree_priors = beautier::create_yule_tree_priors(
-    beautier::get_alignment_ids(fasta_filenames)),
+  site_models = list(
+    beautier::create_jc69_site_model(
+      beautier::get_alignment_id(fasta_filenames)
+    )
+  ),
+  clock_models = list(
+    beautier::create_strict_clock_model(
+      beautier::get_alignment_id(fasta_filenames)
+    )
+  ),
+  tree_priors = list(
+    beautier::create_yule_tree_prior(
+      beautier::get_alignment_id(fasta_filenames)
+    )
+  ),
   mrca_priors = NA,
   mcmc = beautier::create_mcmc(),
   posterior_crown_age = NA,
