### 2018-11-13 #209 @bjoelle
@richelbilderbeek @noamross I have reviewed the changes. One additional thing I would suggest would be to include the overwrite parameter in ``babette::bbt_run`` (it's only present in ``beastier::run_beast2`` at present), and to set it to FALSE by default (as it is in BEAST2).
Otherwise I am happy with the state of the package and would recommend its approval.
