### 2018-11-15 0fcf0c5
Separated 'bbt_run' files for normal and Nested Sampling runs
