### 2018-11-15 de27106
A nested sampling run runs until convergence
