### 2018-11-15 4b773b1
NS has its output parsed, expose #37
