### 2018-11-15 e5cdfe9
Nested Sampling run leaves no temporaries, fix #36
