### 2018-11-16 26f18fc
Fix AppVeyor error
