+++ man/bbt_process_pkg_output.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bbt_process_pkg_output.R
-\name{bbt_process_pkg_output}
-\alias{bbt_process_pkg_output}
-\title{Process the BEAST2 output dependent on BEAST2 package specifics}
-\usage{
-bbt_process_pkg_output(out, mcmc, alignment_ids = NA)
-}
-\arguments{
-\item{out}{complete babette output}
-
-\item{mcmc}{an MCMC}
-}
-\description{
-Process the BEAST2 output dependent on BEAST2 package specifics
-}
