### 2018-11-16 18d152a
Thanks @lintr-bot
