### 2018-11-16 a9309b9
Update pictures, add orange-fronted barbet
