### 2018-11-16 288ce1e
Run tests on Travis and AppVeyor
