### 2018-11-19 richelbilderbeek/babette#38 @richelbilderbeek
Add overwrite parameter to bbt_run
From [here](https://github.com/ropensci/onboarding/issues/209#issuecomment-438320962):

>  One additional thing I would suggest would be to include the overwrite parameter in babette::bbt_run (it's only present in beastier::run_beast2 at present), and to set it to FALSE by default (as it is in BEAST2).
