### 2018-11-21 aa94ca9
Improve documentation about marginal likelihood estimation using Nested Sampling
