### 2018-11-29 #209 @dwinter
@richelbilderbeek @noamross ,

I have a chance now to check out the changes to all these packages. I want to say @richelbilderbeek should be commended for the work he has done to take our comments on board and improve the packages. I particularly like the way he has used the issue tracker to explain the rationale for changes that don't' exactly match the reviewer's suggestions.

I think this is an excellent suite of packages that will support reproducible approaches in an important field, so I'm happy to recommend it be approved.
