### 2018-11-30 f0d396d
Travis assumes Java is installed
