### 2018-11-30 912f094
Simplify Travis script
