### 2018-11-30 d9a1ecd
Add tip dating
