### 2018-11-30 71627e8
Use tree prior that can handle tip dating
