### 2018-11-30 4edb3bd
Actually run babette on tip dating
