### 2018-11-30 1a2d378
Thanks @lintr-bot
