### 2018-12-03 #209 @richelbilderbeek
Hi @noamross,

I am in a dilemma here, as I want to remove some half-baked functionality. This will change the interface a lot:

```bbt_run <- function(
  fasta_filename, # Was 'fasta_filenames' (plural)
  tipdates_filename = NA, # Added one week ago
  site_model, # Was 'site_models' (plural)
  clock_model, # Was 'clock_models' (plural)
  tree_prior, # Was 'tree_priors' (plural)
  # posterior_crown_age = NA, # Remove
)
```

If rOpenSci is flexible in allowing me to change the interface, I will first round my to-dos. After that, I will modify the interface as above, before I submit the package to CRAN.

Is that the preferred way to go? If no, what is?
