### 2018-12-03 #209 @noamross
These changes are fine, and I suggest you make the changes, transfer, set the version, then and submit to CRAN with the new URLs in DESCRIPTION, so that the accepted version is the same as what is on CRAN.  (We are started to roll out versioned peer-review badges so that it is clear what the version was at acceptance.)  You are free to continue to develop and make both small and big changes after acceptance, we have a section of our dev guide dedicated to ongoing changes and maintenance: https://ropensci.github.io/dev_guide/evolution.html
