### 2018-12-07 b7484fb
Remove posterior_crown_age from babette, #39
