+++ tests/testthat/test-bbt_run.R
@@ -5,42 +5,42 @@ test_that("use, one alignment", {
 
   out <- NA
 
-  testthat::expect_silent(
+  expect_silent(
     out <- bbt_run(
       fasta_filenames = get_babette_path("anthus_aco.fas"),
       mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
     )
   )
-  testthat::expect_true("estimates" %in% names(out))
-  testthat::expect_true("anthus_aco_trees" %in% names(out))
-  testthat::expect_true("operators" %in% names(out))
-  testthat::expect_true("output" %in% names(out))
-  testthat::expect_equal(class(out$anthus_aco_trees[[1]]), "phylo")
-  testthat::expect_equal(length(out$anthus_aco_trees), 2)
+  expect_true("estimates" %in% names(out))
+  expect_true("anthus_aco_trees" %in% names(out))
+  expect_true("operators" %in% names(out))
+  expect_true("output" %in% names(out))
+  expect_equal(class(out$anthus_aco_trees[[1]]), "phylo")
+  expect_equal(length(out$anthus_aco_trees), 2)
 
-  testthat::expect_true("Sample" %in% names(out$estimates))
-  testthat::expect_true("posterior" %in% names(out$estimates))
-  testthat::expect_true("likelihood" %in% names(out$estimates))
-  testthat::expect_true("prior" %in% names(out$estimates))
-  testthat::expect_true("treeLikelihood" %in% names(out$estimates))
-  testthat::expect_true("TreeHeight" %in% names(out$estimates))
-  testthat::expect_true("YuleModel" %in% names(out$estimates))
-  testthat::expect_true("birthRate" %in% names(out$estimates))
+  expect_true("Sample" %in% names(out$estimates))
+  expect_true("posterior" %in% names(out$estimates))
+  expect_true("likelihood" %in% names(out$estimates))
+  expect_true("prior" %in% names(out$estimates))
+  expect_true("treeLikelihood" %in% names(out$estimates))
+  expect_true("TreeHeight" %in% names(out$estimates))
+  expect_true("YuleModel" %in% names(out$estimates))
+  expect_true("birthRate" %in% names(out$estimates))
 
-  testthat::expect_true("operator" %in% names(out$operators))
-  testthat::expect_true("p" %in% names(out$operators))
-  testthat::expect_true("accept" %in% names(out$operators))
-  testthat::expect_true("reject" %in% names(out$operators))
-  testthat::expect_true("acceptFC" %in% names(out$operators))
-  testthat::expect_true("rejectFC" %in% names(out$operators))
-  testthat::expect_true("rejectIv" %in% names(out$operators))
-  testthat::expect_true("rejectOp" %in% names(out$operators))
+  expect_true("operator" %in% names(out$operators))
+  expect_true("p" %in% names(out$operators))
+  expect_true("accept" %in% names(out$operators))
+  expect_true("reject" %in% names(out$operators))
+  expect_true("acceptFC" %in% names(out$operators))
+  expect_true("rejectFC" %in% names(out$operators))
+  expect_true("rejectIv" %in% names(out$operators))
+  expect_true("rejectOp" %in% names(out$operators))
 })
 
 test_that("use, one alignment, verbose, cleanup", {
   testit::assert(beastier::is_beast2_installed())
 
-  testthat::expect_output(
+  expect_output(
     bbt_run(
       fasta_filenames = get_babette_path("anthus_aco.fas"),
       mcmc = create_mcmc(chain_length = 1000, store_every = 1000),
[…]
