### 2018-12-07 429663f
Remove posterior_crown_age from babette, #39
