+++ man/bbt_run.Rd
@@ -24,7 +24,6 @@ bbt_run(
   ),
   mrca_priors = NA,
   mcmc = beautier::create_mcmc(),
-  posterior_crown_age = NA,
   beast2_input_filename = tempfile(pattern = "beast2_", fileext = ".xml"),
   rng_seed = 1,
   beast2_output_log_filename = tempfile(
@@ -41,6 +40,7 @@ bbt_run(
     pattern = "beast2_", fileext = ".xml.state"
   ),
   beast2_path = beastier::get_default_beast2_path(),
+  overwrite = FALSE,
   verbose = FALSE,
   cleanup = TRUE
 )
