### 2018-12-07 878a334
Remove posterior_crown_age from babette, #39
