### 2018-12-07 5d1d41c
Properly remove posterior_crown_age from babette, #39
