+++ R/bbt_run.R
@@ -1,6 +1,7 @@
 #' Do a full run: create a BEAST2 configuration file (like BEAUti 2),
 #' run BEAST2, parse results (like Tracer)
 #' @inheritParams bbt_default_params_doc
+#' @param posterior_crown_age deprecated
 #' @return a list with the following elements:\cr
 #' \itemize{
 #'   \item{
@@ -78,7 +79,8 @@
 #'   beast2_path = beastier::get_default_beast2_path(),
 #'   overwrite = FALSE,
 #'   verbose = FALSE,
-#'   cleanup = TRUE
+#'   cleanup = TRUE,
+#'  posterior_crown_age = NA
 #' )
 #' @examples
 #'  out <- bbt_run(
@@ -150,7 +152,8 @@ bbt_run <- function(
   beast2_path = beastier::get_default_beast2_path(),
   overwrite = FALSE,
   verbose = FALSE,
-  cleanup = TRUE
+  cleanup = TRUE,
+  posterior_crown_age = NA
 ) {
 
   # Check for deprecated argument names
