### 2018-12-07 a23214f
Support one MRCA prior, #40
