### 2018-12-07 c779793
Support one tree prior, #41
