### 2018-12-07 c413bc7
Support one tree prior, #41
