+++ R/bbt_default_params_doc.R
@@ -53,11 +53,14 @@
 #'     \item{the \code{.log} file created by BEAST2 exists}
 #'     \item{the \code{.trees} files created by BEAST2 exist}
 #'  }
+#' @param posterior_crown_age deprecated
 #' @param rng_seed the random number generator seed. Must be either
 #'   \code{NA} or a positive non-zero value. An RNG seed of \code{NA}
 #'   results in BEAST2 picking a random seed.
 #' @param site_models one or more site models,
 #'   see \link[beautier]{create_site_models}
+#' @param tree_prior one tree priors,
+#'   as created by \link[beautier]{create_tree_prior}
 #' @param tree_priors one or more tree priors,
 #'   see \link[beautier]{create_tree_priors}
 #' @param verbose set to TRUE for more output
@@ -78,8 +81,10 @@ bbt_default_params_doc <- function(
   mrca_prior,
   mrca_priors,
   overwrite,
+  posterior_crown_age,
   rng_seed,
   site_models,
+  tree_prior,
   tree_priors,
   verbose
 ) {
