+++ man/bbt_default_params_doc.Rd
@@ -7,9 +7,9 @@ documentation.}
 \usage{
 bbt_default_params_doc(beast2_input_filename, beast2_output_log_filename,
   beast2_output_state_filename, beast2_output_trees_filenames, beast2_path,
-  cleanup, clock_models, fasta_filenames, mcmc, mrca_prior, mrca_priors,
-  overwrite, posterior_crown_age, rng_seed, site_models, tree_prior,
-  tree_priors, verbose)
+  cleanup, clock_model, clock_models, fasta_filenames, mcmc, mrca_prior,
+  mrca_priors, overwrite, posterior_crown_age, rng_seed, site_models,
+  tree_prior, tree_priors, verbose)
 }
 \arguments{
 \item{beast2_input_filename}{path of the BEAST2 configuration file.
@@ -52,6 +52,9 @@ Use \code{get_default_beast2_jar_path} to get
 the default BEAST jar file's path}
 
 \item{cleanup}{set to FALSE to keep all temporary files}
+
+\item{clock_model}{one clock model,
+see \link[beautier]{create_clock_model}}
 
 \item{clock_models}{one or more clock models,
 see \link[beautier]{create_clock_models}}

[…]
