+++ vignettes/article.R
@@ -15,51 +15,3 @@ file.copy(babette::get_babette_path("anthus_nd2_sub.fas"), "anthus_nd2.fas")
 library(babette)
 mcmc <- create_mcmc(chain_length = 2000, store_every = 1000)
 
-## ------------------------------------------------------------------------
-out <- bbt_run(
-  fasta_filenames = "anthus_aco.fas", 
-  mcmc = mcmc
-)
-
-## ------------------------------------------------------------------------
-out <- bbt_run(
-  "anthus_aco.fas",
-  site_models = create_hky_site_model(),
-  clock_model = create_rln_clock_model(),
-  tree_prior = create_bd_tree_prior(),
-  mcmc = mcmc
-)
-
-## ------------------------------------------------------------------------
-# Currently, support for multiple alignments has been deprecated
-if (1 == 2) {
-  out <- bbt_run(
-    c("anthus_aco.fas", "anthus_nd2.fas"),
-    site_models = list(
-      create_tn93_site_model(), 
-      create_gtr_site_model()
-    ),
-    mcmc = mcmc
-  )
-}
-
-## ------------------------------------------------------------------------
-out <- bbt_run(
-  "anthus_aco.fas",
-  tree_prior = create_yule_tree_prior(
-    birth_rate_distr = create_exp_distr()    
-  ),
-  mcmc = mcmc
-)
-
-## ------------------------------------------------------------------------
-out <- bbt_run(
-  "anthus_aco.fas",
-  tree_prior = create_yule_tree_prior(
-    birth_rate_distr = create_exp_distr(
-      mean = 1.0
-    )    
-  ),
-  mcmc = mcmc
-)
-
