+++ vignettes/tutorial.R
@@ -22,16 +22,16 @@ site_model <- create_jc69_site_model()
 clock_model <- create_clock_model_strict()
 clock_model <- create_strict_clock_model()
 
-## ------------------------------------------------------------------------
-tree_prior <- create_tree_prior_yule()
-tree_prior <- create_yule_tree_prior()
-
 ## ----cache=TRUE----------------------------------------------------------
 out <- bbt_run(
   fasta_filenames = fasta_filename,
-  tree_prior = tree_prior,
+  clock_model = clock_model,
   mcmc = mcmc
 )
+
+## ------------------------------------------------------------------------
+tree_prior <- create_tree_prior_yule()
+tree_prior <- create_yule_tree_prior()
 
 ## ------------------------------------------------------------------------
 mrca_prior <- create_mrca_prior(
