+++ man/bbt_default_params_doc.Rd
@@ -8,8 +8,8 @@ documentation.}
 bbt_default_params_doc(beast2_input_filename, beast2_output_log_filename,
   beast2_output_state_filename, beast2_output_trees_filenames, beast2_path,
   cleanup, clock_model, clock_models, fasta_filenames, mcmc, mrca_prior,
-  mrca_priors, overwrite, posterior_crown_age, rng_seed, site_models,
-  tree_prior, tree_priors, verbose)
+  mrca_priors, overwrite, posterior_crown_age, rng_seed, site_model,
+  site_models, tree_prior, tree_priors, verbose)
 }
 \arguments{
 \item{beast2_input_filename}{path of the BEAST2 configuration file.
@@ -85,6 +85,9 @@ as returned by \code{\link{create_mrca_prior}}}
 \item{rng_seed}{the random number generator seed. Must be either
 \code{NA} or a positive non-zero value. An RNG seed of \code{NA}
 results in BEAST2 picking a random seed.}
+
+\item{site_model}{one site model,
+see \link[beautier]{create_site_models}}
 
 \item{site_models}{one or more site models,
 see \link[beautier]{create_site_models}}
