### 2018-12-07 d650618
Support one FASTA file, #44
