+++ tests/testthat/test-bbt_run.R
@@ -7,7 +7,7 @@ test_that("use, one alignment", {
 
   expect_silent(
     out <- bbt_run(
-      fasta_filenames = get_babette_path("anthus_aco.fas"),
+      fasta_filename = get_babette_path("anthus_aco.fas"),
       mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
     )
   )
@@ -42,7 +42,7 @@ test_that("use, one alignment, verbose, cleanup", {
 
   expect_output(
     bbt_run(
-      fasta_filenames = get_babette_path("anthus_aco.fas"),
+      fasta_filename = get_babette_path("anthus_aco.fas"),
       mcmc = create_mcmc(chain_length = 1000, store_every = 1000),
       verbose = TRUE
     )
@@ -67,7 +67,7 @@ test_that("use, one alignment, verbose, no cleanup", {
 
   expect_output(
     bbt_run(
-      fasta_filenames = fasta_filenames,
+      fasta_filename = fasta_filenames,
       mcmc = create_mcmc(chain_length = 1000, store_every = 1000),
       beast2_input_filename = beast2_input_filename,
       beast2_output_log_filename = beast2_output_log_filename,
@@ -93,12 +93,12 @@ test_that("use, one alignment, same RNG should give same results", {
 
   rng_seed <- 42
   out_1 <- bbt_run(
-    fasta_filenames = get_babette_path("anthus_aco.fas"),
+    fasta_filename = get_babette_path("anthus_aco.fas"),
     mcmc = create_mcmc(chain_length = 1000, store_every = 1000),
     rng_seed = rng_seed
   )
   out_2 <- bbt_run(
-    fasta_filenames = get_babette_path("anthus_aco.fas"),
+    fasta_filename = get_babette_path("anthus_aco.fas"),
     mcmc = create_mcmc(chain_length = 1000, store_every = 1000),
     rng_seed = rng_seed
   )
@@ -116,7 +116,7 @@ test_that("abuse", {
 
   expect_error(
     bbt_run(
-      fasta_filenames = get_babette_path("anthus_aco.fas"),
+      fasta_filename = get_babette_path("anthus_aco.fas"),
       beast2_output_trees_filenames = c("too", "many")
     ),
     "Must have as much FASTA filenames as BEAST2 output trees filenames"
@@ -132,7 +132,7 @@ test_that("Run all defaults", {
   if (!beastier::is_on_ci()) return()
   expect_silent(
     bbt_run(
-      fasta_filenames = get_beautier_path("anthus_aco_sub.fas"),
+      fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
       mcmc = beautier::create_mcmc(chain_length = 2000)
     )
   )
@@ -150,7 +150,7 @@ test_that("Run GTR", {
   if (!beastier::is_on_ci()) return()
   expect_silent(
     bbt_run(
-      fasta_filenames = get_beautier_path("anthus_aco_sub.fas"),
+      fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
       site_model = create_gtr_site_model(),
       mcmc = beautier::create_mcmc(chain_length = 2000)
     )
@@ -165,7 +165,7 @@ test_that("Run HKY", {
   if (!beastier::is_on_ci()) return()
   expect_silent(
     bbt_run(
-      fasta_filenames = get_beautier_path("anthus_aco_sub.fas"),
+      fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
       site_model = create_hky_site_model(),
       mcmc = beautier::create_mcmc(chain_length = 2000)
     )
@@ -180,7 +180,7 @@ test_that("Run JC69", {
[…]
