+++ vignettes/nested_sampling.R
@@ -76,13 +76,13 @@ mcmc <- beautier::create_mcmc_nested_sampling(
 ## ------------------------------------------------------------------------
 out_jc69 <- bbt_run(
   fasta_filenames = fasta_filename,
-  site_models = beautier::create_jc69_site_model(),
+  site_model = beautier::create_jc69_site_model(),
   mcmc = mcmc,
   beast2_path = beastier::get_default_beast2_bin_path()
 )
 out_gtr <- bbt_run(
   fasta_filenames = fasta_filename,
-  site_models = beautier::create_gtr_site_model(),
+  site_model = beautier::create_gtr_site_model(),
   mcmc = mcmc,
   beast2_path = beastier::get_default_beast2_bin_path()
 )
