### 2018-12-07 baf0acb
Support one FASTA file, #44
