+++ vignettes/tutorial.R
@@ -14,13 +14,19 @@ testit::assert(file.exists(fasta_filename))
 ## ------------------------------------------------------------------------
 mcmc <- create_mcmc(chain_length = 2000, store_every = 1000)
 
+## ----cache=TRUE----------------------------------------------------------
+out <- bbt_run(
+ fasta_filename = fasta_filename,
+ mcmc = mcmc
+)
+
 ## ------------------------------------------------------------------------
 site_model <- create_site_model_jc69()
 site_model <- create_jc69_site_model()
 
 ## ----cache=TRUE----------------------------------------------------------
 out <- bbt_run(
-  fasta_filenames = fasta_filename,
+  fasta_filename = fasta_filename,
   site_model = site_model,
   mcmc = mcmc
 )
@@ -29,9 +35,23 @@ out <- bbt_run(
 clock_model <- create_clock_model_strict()
 clock_model <- create_strict_clock_model()
 
+## ----cache=TRUE----------------------------------------------------------
+out <- bbt_run(
+  fasta_filename = fasta_filename,
+  clock_model = clock_model,
+  mcmc = mcmc
+)
+
 ## ------------------------------------------------------------------------
 tree_prior <- create_tree_prior_yule()
 tree_prior <- create_yule_tree_prior()
+
+## ----cache=TRUE----------------------------------------------------------
+out <- bbt_run(
+  fasta_filename = fasta_filename,
+  tree_prior = tree_prior,
+  mcmc = mcmc
+)
 
 ## ------------------------------------------------------------------------
 mrca_prior <- create_mrca_prior(
@@ -53,6 +73,28 @@ mrca_prior <- create_mrca_prior(
   mrca_distr = mrca_distr
 )
 
+## ----cache=TRUE----------------------------------------------------------
+out <- bbt_run(
+  fasta_filename = fasta_filename,
+  mrca_prior = mrca_prior,
+  mcmc = mcmc
+)
+
+## ----cache=TRUE----------------------------------------------------------
+# Prefer this over using 'posterior_crown_age'
+out <- bbt_run(
+  fasta_filename = fasta_filename,
+  mrca_prior = create_mrca_prior(
+    alignment_id = get_alignment_id(fasta_filename = fasta_filename), 
+    taxa_names = get_taxa_names(filename = fasta_filename),
+    mrca_distr = create_normal_distr(
+      mean = 15.0, 
+      sigma = 0.0001
+    )
+  ),
+  mcmc = mcmc
+)
+
 ## ------------------------------------------------------------------------
 if (1 == 2) {
   beast2_input_filename <- "beast_input.xml"
@@ -67,7 +109,7 @@ if (1 == 2) {
   )
     
   out <- bbt_run(
-    fasta_filenames = fasta_filename,
+    fasta_filename = fasta_filename,
     mcmc = mcmc,
     beast2_input_filename = beast2_input_filename,
     beast2_output_log_filename = beast2_output_log_filename,
@@ -78,6 +120,13 @@ if (1 == 2) {
   file.remove(all_files)
 }
 
+## ----cache=TRUE----------------------------------------------------------
+out <- bbt_run(
+  fasta_filename = fasta_filename,
+  mcmc = mcmc,
+  rng_seed = 314
+)
+
 ## ------------------------------------------------------------------------
 beast2_path <- beastier::get_default_beast2_path()
 print(beast2_path)
@@ -85,9 +134,16 @@ print(beast2_path)
 ## ------------------------------------------------------------------------
 if (file.exists(beast2_path)) {
   out <- bbt_run(
-    fasta_filenames = fasta_filename,
+    fasta_filename = fasta_filename,
     mcmc = mcmc,
[…]
