### 2018-12-07 e2599ac
Cannot have different amount of (one) FASTA filename and (one) output filenames
