### 2018-12-07 9c41613
Cannot have different amount of (one) FASTA filename and (one) output filenames
