### 2018-12-07 f553497
Thanks @lintr-bot
