### 2018-12-10 dc31d9e
Follow new interface 'beautier'
