+++ R/bbt_run.R
@@ -43,6 +43,7 @@
 #' @usage
 #' bbt_run(
 #'   fasta_filename,
+#'   tipdates_filename = NA,
 #'   site_model = beautier::create_jc69_site_model(),
 #'   clock_model = beautier::create_strict_clock_model(),
 #'   tree_prior = beautier::create_yule_tree_prior(),
