### 2018-12-18 c5b31f1
phangorn::densiTree must be silent for one phylogeny
