### 2018-12-19 21a4d5d
Rely on devel branch of phangorn
