+++ tests/testthat/test-plot_densitree.R
@@ -12,6 +12,7 @@ test_that("use", {
 })
 
 test_that("use", {
+  skip("Do not test phangorn")
   phylogeny <- ape::read.tree(text = "(((A:1, B:1):1, C:2):1, D:3);")
   expect_silent(plot_densitree(c(phylogeny)))
 })
