+++ tests/testthat/test-bbt_check_beast2_packages.R
@@ -2,6 +2,12 @@ context("bbt_check_beast2_packages")
 
 test_that("use", {
 
+  # This test uninstalls the NS BEAST2 package.
+  # Only do that on CI services, else a user without internet
+  # suddenly finds the NS BEAST2 package installed and unable
+  # to reinstall it
+  if (!beastier::is_on_ci()) return()
+
   beast2_package_name <- "NS"
   # Check to need to install NS later
   was_ns_installed <- mauricer::mrc_is_installed(beast2_package_name)
