+++ tests/testthat/test-bbt_run_ns.R
@@ -72,7 +72,13 @@ test_that("Nested sampling run should create no temporaries", {
   old_work_dir <- getwd()
   new_work_dir <- tempdir()
   setwd(new_work_dir)
-  n_files_before <- length(list.files(new_work_dir))
+  files_before <- list.files(new_work_dir)
+
+  # Temporary files created
+  beast2_input_filename <- tempfile(".xml")
+  beast2_output_log_filename <- tempfile("log")
+  beast2_output_trees_filenames <- tempfile("trees")
+  beast2_output_state_filename <- tempfile(".state.xml")
 
   bbt_run(
     fasta_filename = get_babette_path("anthus_aco.fas"),
@@ -82,14 +88,23 @@ test_that("Nested sampling run should create no temporaries", {
       sub_chain_length = 500,
       epsilon = 1.0
     ),
-    beast2_path = get_default_beast2_bin_path()
+    beast2_path = get_default_beast2_bin_path(),
+    beast2_input_filename = beast2_input_filename,
+    beast2_output_log_filename = beast2_output_log_filename,
+    beast2_output_trees_filenames = beast2_output_trees_filenames,
+    beast2_output_state_filename = beast2_output_state_filename
   )
 
-  n_files_after <- length(list.files(new_work_dir))
+  # Delete the temporary files
+  file.remove(beast2_input_filename)
+  file.remove(beast2_output_log_filename)
+  file.remove(beast2_output_trees_filenames)
+  file.remove(beast2_output_state_filename)
+
+  files_after <- list.files(new_work_dir)
   setwd(old_work_dir)
   list.files(new_work_dir)
-  expect_equal(n_files_before, n_files_after)
-
+  expect_equal(files_before, files_after)
 })
 
 test_that("abuse", {
