+++ R/zzz.R
@@ -2,7 +2,7 @@
 
   suppressPackageStartupMessages(
     lapply(
-      c("beautier", "beastier", "tracerer"),
+      c("beautier", "beastier", "tracerer", "mauricer"),
       library,
       character.only = TRUE,
       warn.conflicts = FALSE

[…]
