### 2019-01-04 a073178
Update links to ropensci
