+++ tests/testthat/test-bbt_run_ns.R
@@ -55,7 +55,7 @@ test_that("use", {
   expect_true("trees" %in% names(out$ns))
 
   skip("TODO: measure Nested Sampling ESS. Issue 37, #37")
-  # https://github.com/richelbilderbeek/babette/issues/37
+  # https://github.com/ropensci/babette/issues/37
   expect_true("ess" %in% names(out$ns))
 })
 
@@ -65,7 +65,7 @@ test_that("Nested sampling run should create no temporaries", {
     skip("Cannot run Nested Sampling package from Windows")
   }
 
-  # From https://github.com/richelbilderbeek/babette/issues/36
+  # From https://github.com/ropensci/babette/issues/36
   testit::assert(mauricer::mrc_is_installed("NS"))
 
   # Run babette in a different folder
