+++ tests/testthat/test-bbt_run.R
@@ -423,7 +423,7 @@ test_that("RLN and monophyletic MRCA with distribution, Issue 29, #29", {
 
 test_that("use, one alignment, plot with nLTT", {
 
-  skip("Expose bug, https://github.com/richelbilderbeek/babette/issues/10")
+  skip("Expose bug, https://github.com/ropensci/babette/issues/10")
   out <- bbt_run(
     fasta_filename = get_babette_path("anthus_aco.fas"),
     mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
