+++ doc/examples.md
@@ -1,6 +1,6 @@
 # Examples
 
- * Want something more interactive? [lumier](https://github.com/richelbilderbeek/lumier) 
+ * Want something more interactive? [lumier](https://github.com/ropensci/lumier) 
    is an R Shiny app to help create the R function call needed!
 
 For all examples, do load `babette`:
