+++ appveyor.yml
@@ -24,10 +24,10 @@ environment:
 build_script:
   - travis-tool.sh install_deps
   - R -e "library(beastier); beastier::install_beast2()"
-  - travis-tool.sh install_github richelbilderbeek/tracerer
-  - travis-tool.sh install_github richelbilderbeek/beautier
-  - travis-tool.sh install_github richelbilderbeek/beastier
-  - travis-tool.sh install_github richelbilderbeek/mauricer
+  - travis-tool.sh install_github ropensci/tracerer
+  - travis-tool.sh install_github ropensci/beautier
+  - travis-tool.sh install_github ropensci/beastier
+  - travis-tool.sh install_github ropensci/mauricer
   - travis-tool.sh install_github jimhester/covr
   - travis-tool.sh install_github jimhester/lintr
   - travis-tool.sh install_github MangoTheCat/goodpractice
