### 2019-01-04 ropensci/babette#55 @richelbilderbeek
Process last things lumier
From #48:

 * [x]  Transfer the repos to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  You'll be made admin once you do.

 * [x]  Add the rOpenSci badge to top and footer to the bottom of your READMEs
       "` [![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)`"

 * [x] Should you want to acknowledge your reviewers in your package DESCRIPTION, you can do so by making them `"rev"`-type contributors in the `Authors@R` field (with their consent). More info on this [here](https://ropensci.github.io/dev_guide/building.html#authorship).

This package was not reviewed, as it was created after the review process.

For me:

 * [x] Tag new version
