### 2019-01-04 496a394
Add rOpenSci badges
