+++ tests/testthat/test-bbt_run_ns.R
@@ -65,13 +65,8 @@ test_that("Nested sampling run should create no temporaries", {
     skip("Cannot run Nested Sampling package from Windows")
   }
 
-<<<<<<< HEAD
-  # From https://github.com/richelbilderbeek/babette/issues/36
-  testit::assert(mauricer::is_beast2_pkg_installed("NS"))
-=======
   # From https://github.com/ropensci/babette/issues/36
-  testit::assert(mauricer::mrc_is_installed("NS"))
->>>>>>> master
+  testit::assert(mauricer::is_beast2_pkg_installed("NS"))
 
   # Run babette in a different folder
   old_work_dir <- getwd()
