### 2019-01-04 bef52b0
Ran 'codemetar::write_codemeta()'
