### 2019-01-04 ropensci/babette#52 @richelbilderbeek
v2.0 is born!
