### 2019-01-04 ropensci/babette#51 @richelbilderbeek
[v2.1](https://github.com/ropensci/beautier/releases/tag/v2.1) is there.
