### 2019-01-04 93f7e74
Use 'mauricer::install_beast2_pkg' instead of deprecated 'mauricer::mrc_install'
