### 2019-01-04 3f8ed75
Use 'mauricer::is_beast2_pkg_installed' instead of deprecated 'mauricer::mrc_is_installed'
