### 2019-01-04 #209 @stefaniebutland
Hello @richelbilderbeek and thank you for all your work on having `babette` go through peer review.

> We'd also love a blog post about your packages, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. (https://ropensci.org/blog/).

We would love to help get more eyes on your work. This link - https://ropensci.org/tags/software-peer-review/ - will give you examples of blog posts by authors of peer-reviewed packages so you can get an idea of the style and length you prefer. Technotes are here: https://ropensci.org/technotes/.

Here are some technical and editorial guidelines for contributing a blog post: https://github.com/ropensci/roweb2#contributing-a-blog-post.

If you are interested in contributing a blog post, please let me know what timing might work for you. Happy to answer any questions.
