### 2019-01-06 ropensci/babette#53 @richelbilderbeek
[v2.0](https://github.com/ropensci/babette/releases/tag/v2.0) is born.
