### 2019-01-08 710fe81
Install libgit2 on MacOS
