### 2019-01-10 #209 @richelbilderbeek
@stefaniebutland: I can submit a draft before the Friday 25th of January, 19:00 (Amsterdam time zone, UTC +1). Would that work out?
