+++ DESCRIPTION
@@ -34,6 +34,7 @@ Suggests:
     ape,
     devtools,
     ggplot2,
+    ggtree,
     hunspell,
     knitr,
     lintr,
