+++ .travis.yml
@@ -11,8 +11,11 @@ r_packages:
   - hunspell
   - xml2
 
+bioc_packages:
+  - ggtree
+
 r_github_packages:
-  - GuangchuangYu/ggtree
+  - # GuangchuangYu/ggtree
   - jimhester/covr
   - jimhester/lintr
   - ropensci/beastier
