### 2019-01-23 a7e3b44
Install magrittr, a prerequisite for ggtree
