### 2019-01-23 6c508d7
Try install magritr in AppVeyor, predict to fail
