+++ appveyor.yml
@@ -22,6 +22,7 @@ environment:
     USE_RTOOLS: true
 
 build_script:
+  - travis-tool.sh install magrittr
   - travis-tool.sh install_deps
   - R -e "library(beastier); beastier::install_beast2()"
   - travis-tool.sh install_github GuangchuangYu/ggtree
