### 2019-01-23 34309b1
DESCRIPTION suggests magrittr
