### 2019-01-23 f3fb423
Install magrittr manually
