+++ .travis.yml
@@ -16,7 +16,7 @@ r_packages:
 #   - ggtree
 
 r_github_packages:
-  - GuangchuangYu/ggtree
+  - GuangchuangYu/ggtree@9416b852753df0957f812840b9b5eda74b37d9b3
   - jimhester/covr
   - jimhester/lintr
   - ropensci/beastier
