+++ DESCRIPTION
@@ -38,7 +38,6 @@ Suggests:
     hunspell,
     knitr,
     lintr,
-    magrittr,
     mauricer,
     nLTT,
     rmarkdown,
