### 2019-01-25 8940f46
Use phangorn as default for plot_densitree
