### 2019-01-25 d5c986a
Clean up vignette folder
