### 2019-01-30 d21c236
Test 'bbt_run_from_model' correctly
