+++ R/bbt_run.R
@@ -200,7 +200,12 @@ bbt_run <- function(
   }
 
   if (length(fasta_filenames) != length(beast2_output_trees_filenames)) {
-    stop("Must have as much FASTA filenames as BEAST2 output trees filenames")
+    stop(
+      "Must have as much FASTA filenames as BEAST2 output trees filenames.\n",
+      "Number of FASTA filenames: ", length(fasta_filenames), ".\n",
+      "Number of BEAST2 output filenames: ",
+      length(beast2_output_trees_filenames), ".\n"
+    )
   }
   if (!is.na(rng_seed) && !(rng_seed > 0)) {
     stop("'rng_seed' should be NA or non-zero positive")
