+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(bbt_run)
+export(bbt_run_from_model)
 export(get_babette_path)
 export(get_babette_paths)
 export(plot_densitree)

[…]
