### 2019-01-31 f9e7ecb
'beast2_options' has a default 'input_filename' value
