+++ .travis.yml
@@ -20,7 +20,7 @@ r_github_packages:
   - jimhester/lintr
   - ropensci/beautier
   - ropensci/tracerer
-  - ropensci/beastier
+  - ropensci/beastier@develop
   - ropensci/mauricer
   - MangoTheCat/goodpractice
   - KlausVigo/phangorn
