+++ .travis.yml
@@ -20,7 +20,7 @@ r_github_packages:
   - jimhester/lintr
   - ropensci/beautier
   - ropensci/tracerer
-  - ropensci/beastier@develop
+  - ropensci/beastier
   - ropensci/mauricer
   - MangoTheCat/goodpractice
   - KlausVigo/phangorn
