### 2019-02-01 889088e
Add rOpenSci blog post figure creation script
