### 2019-02-01 db183f4
Thanks @lintr-bot
