### 2019-02-01 58dd811
phangorn bug fixed, thanks @KlausVigo
