+++ tests/testthat/test-plot_densitree.R
@@ -18,9 +18,12 @@ test_that("use", {
   )
 })
 
-test_that("use", {
-  skip("Do not test phangorn")
-  trees <- c(ape::read.tree(text = "(((A:1, B:1):1, C:2):1, D:3);"))
+test_that("minimal use", {
+  # Submitted at https://github.com/KlausVigo/phangorn/commit/8440b88fb09249a2653562f0eaee04a51789bcba # nolint indeed URLs can be long
+  phylogeny_1 <- ape::read.tree(text = "((A:2, B:2):1, C:3);")
+  phylogeny_2 <- ape::read.tree(text = "((A:1, B:1):2, C:3);")
+  trees <- c(phylogeny_1, phylogeny_2)
+  expect_silent(phangorn::densiTree(trees))
   expect_silent(plot_densitree(trees))
 })
 
