### 2019-02-04 c4bc392
nested sampling works (and worked) great with 'bbt_run_from_model', fix #64
