+++ tests/testthat/test-bbt_run_from_model.R
@@ -56,3 +56,19 @@ test_that("abuse", {
     "'beast2_options' must be a valid BEAST2 options object"
   )
 })
+
+test_that("use, one alignment", {
+  # https://github.com/richelbilderbeek/pirouette/issues/99
+  testit::assert(beastier::is_beast2_installed())
+
+  bbt_out <- bbt_run_from_model(
+    fasta_filename = get_babette_path("anthus_aco.fas"),
+    inference_model = create_inference_model(
+      tree_prior = create_yule_tree_prior(),
+      mcmc = create_mcmc(chain_length = 3000, store_every = 1000)
+    ),
+    beast2_options = create_beast2_options(rng_seed = 42)
+  )
+  expect_equal(4, length(bbt_out$anthus_aco_trees))
+  expect_equal(4, nrow(bbt_out$estimates))
+})
