+++ R/bbt_default_params_doc.R
@@ -74,7 +74,7 @@
 #' @param tree_priors one or more tree priors,
 #'   see \link[beautier]{create_tree_priors}
 #' @param verbose set to TRUE for more output
-#' @author Richel J.C. Bilderbeek
+#' @author Richèl J.C. Bilderbeek
 #' @note This is an internal function, so it should be marked with
 #'   \code{@noRd}. This is not done, as this will disallow all
 #'   functions to find the documentation parameters
