+++ R/plot_densitree.R
@@ -12,7 +12,7 @@
 #' @param ... options to be passed to \code{phangorn}'s
 #'   \link[phangorn]{densiTree} function
 #' @return nothing. Will produce a plot.
-#' @author Richel J.C. Bilderbeek
+#' @author Richèl J.C. Bilderbeek
 #' @examples
 #'   out <- bbt_run(
 #'    get_babette_path("anthus_aco.fas"),
