+++ R/bbt_get_paths.R
@@ -2,7 +2,7 @@
 #' @param filenames the files' names, without the path
 #' @return the filenames' full paths, if and only if
 #'   all files are present. Will stop otherwise.
-#' @author Richel J.C. Bilderbeek
+#' @author Richèl J.C. Bilderbeek
 #' @seealso for one file, use \code{\link{get_babette_path}}
 #' @examples
 #'   testit::assert(
