+++ appveyor.yml
@@ -25,7 +25,6 @@ build_script:
   - # travis-tool.sh install magrittr
   - travis-tool.sh install_deps
   - R -e "library(beastier); beastier::install_beast2()"
-  - travis-tool.sh install_github GuangchuangYu/ggtree
   - travis-tool.sh install_github ropensci/tracerer
   - travis-tool.sh install_github ropensci/beautier
   - travis-tool.sh install_github ropensci/beastier
