### 2019-02-19 5e70fbd
Progress #65. Confirmed to work on Linux, should fail on MacOS
