### 2019-02-19 9ba573c
Add test for #65
