### 2019-02-22 3bdedc7
Use short test for #65
