+++ tests/testthat/test-bbt_run.R
@@ -533,7 +533,7 @@ test_that("file.exists(output_log_filename) is not TRUE, on MacOS, #65", {
 
   posterior <- bbt_run(
     fasta_filename,
-    mcmc = create_mcmc(chain_length = 3000, store_every = 1000),
+    mcmc = create_mcmc(chain_length = 1000000, store_every = 5000),
     rng_seed = 42,
     verbose = TRUE,
     overwrite = TRUE
