### 2019-02-22 4774635
Add complete test for #65
