### 2019-02-22 b07f981
Better assert
