+++ DESCRIPTION
@@ -41,7 +41,6 @@ Suggests:
     nLTT,
     rmarkdown,
     spelling,
-    TESS,
     testthat
 Language: en-US
 Encoding: UTF-8
