### 2019-02-22 1d95ffc
Build for 3 R versions
