+++ .travis.yml
@@ -1,4 +1,9 @@
 language: r
+r:
+  - oldrel
+  - release
+  - devel
+
 cache: packages
 
 os: 
