### 2019-03-07 7ad8476
Fix Travis MacOS build error: 'clang: error: unsupported option '-fopenmp''. thank https://stackoverflow.com/questions/46274159/installation-issue-clang-error-unsupported-option-fopenmp-macos-sierra#comment79510841_46274159
