### 2019-03-07 cb7904d
babette installs clang on Travis
