### 2019-03-26 5bdafe9
Travis does not brew clang, use R binary package of phangorn
