### 2019-03-27 5d556d3
Builds without BEAST2 installed, like CRAN
