+++ tests/testthat/test-bbt_run.R
@@ -1,6 +1,9 @@
 context("bbt_run")
 
 test_that("use, one alignment", {
+
+  if (!beastier::is_beast2_installed()) return()
+
   testit::assert(beastier::is_beast2_installed())
 
   out <- NA
@@ -39,6 +42,9 @@ test_that("use, one alignment", {
 })
 
 test_that("use, one alignment, verbose, cleanup", {
+
+  if (!beastier::is_beast2_installed()) return()
+
   testit::assert(beastier::is_beast2_installed())
 
   expect_output(
@@ -51,6 +57,9 @@ test_that("use, one alignment, verbose, cleanup", {
 })
 
 test_that("use, one alignment, verbose, no cleanup", {
+
+  if (!beastier::is_beast2_installed()) return()
+
   testit::assert(beastier::is_beast2_installed())
 
   fasta_filenames <- get_babette_path("anthus_aco.fas")
@@ -84,6 +93,8 @@ test_that("use, one alignment, verbose, no cleanup", {
 })
 
 test_that("use, one alignment, same RNG should give same results", {
+
+  if (!beastier::is_beast2_installed()) return()
 
   if (rappdirs::app_dir()$os == "mac") {
     skip("RNG seed is unreliable under macOS")
@@ -129,7 +140,10 @@ test_that("abuse", {
 ################################################################################
 
 test_that("Run all defaults", {
+
+  if (!beastier::is_beast2_installed()) return()
   if (!beastier::is_on_ci()) return()
+
   expect_silent(
     bbt_run(
       fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
@@ -147,7 +161,10 @@ test_that("Run all defaults", {
 ################################################################################
 
 test_that("Run GTR", {
+
   if (!beastier::is_on_ci()) return()
+  if (!beastier::is_beast2_installed()) return()
+
   expect_silent(
     bbt_run(
       fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
@@ -162,7 +179,10 @@ test_that("Run GTR", {
 ################################################################################
 
 test_that("Run HKY", {
+
   if (!beastier::is_on_ci()) return()
+  if (!beastier::is_beast2_installed()) return()
+
   expect_silent(
     bbt_run(
       fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
@@ -177,7 +197,10 @@ test_that("Run HKY", {
 ################################################################################
 
 test_that("Run JC69", {
+
   if (!beastier::is_on_ci()) return()
+  if (!beastier::is_beast2_installed()) return()
+
   expect_silent(
     bbt_run(
       fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
@@ -192,7 +215,10 @@ test_that("Run JC69", {
 ################################################################################
 
 test_that("Run TN93", {
+
   if (!beastier::is_on_ci()) return()
+  if (!beastier::is_beast2_installed()) return()
+
   expect_silent(
     bbt_run(
       fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
@@ -210,7 +236,10 @@ test_that("Run TN93", {
 # Clock model: RLN
 ################################################################################
 test_that("Run RLN clock", {
+
   if (!beastier::is_on_ci()) return()
+  if (!beastier::is_beast2_installed()) return()
+
   expect_silent(
     bbt_run(
       fasta_filename = get_beautier_path("anthus_aco_sub.fas"),
@@ -224,7 +253,10 @@ test_that("Run RLN clock", {
 # Clock model: strict
 ################################################################################
[…]
