### 2019-03-27 5bd02a4
Add 'get_bbt_run_test_output'
