+++ R/bbt_run.R
@@ -81,30 +81,31 @@
 #'    mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
 #'  )
 #'
-#'  testit::assert("estimates" %in% names(out))
-#'  testit::assert("anthus_aco_trees" %in% names(out))
-#'  testit::assert("operators" %in% names(out))
-#'  testit::assert("output" %in% names(out))
-#'  testit::assert(class(out$anthus_aco_trees[[1]]) == "phylo")
-#'  testit::assert(length(out$anthus_aco_trees) == 2)
+#'  library(testthat)
+#'  expect_true("estimates" %in% names(out))
+#'  expect_true("anthus_aco_trees" %in% names(out))
+#'  expect_true("operators" %in% names(out))
+#'  expect_true("output" %in% names(out))
+#'  expect_true(class(out$anthus_aco_trees[[1]]) == "phylo")
+#'  expect_true(length(out$anthus_aco_trees) == 2)
 #'
-#'  testit::assert("Sample" %in% names(out$estimates))
-#'  testit::assert("posterior" %in% names(out$estimates))
-#'  testit::assert("likelihood" %in% names(out$estimates))
-#'  testit::assert("prior" %in% names(out$estimates))
-#'  testit::assert("treeLikelihood" %in% names(out$estimates))
-#'  testit::assert("TreeHeight" %in% names(out$estimates))
-#'  testit::assert("YuleModel" %in% names(out$estimates))
-#'  testit::assert("birthRate" %in% names(out$estimates))
+#'  expect_true("Sample" %in% names(out$estimates))
+#'  expect_true("posterior" %in% names(out$estimates))
+#'  expect_true("likelihood" %in% names(out$estimates))
+#'  expect_true("prior" %in% names(out$estimates))
+#'  expect_true("treeLikelihood" %in% names(out$estimates))
+#'  expect_true("TreeHeight" %in% names(out$estimates))
+#'  expect_true("YuleModel" %in% names(out$estimates))
+#'  expect_true("birthRate" %in% names(out$estimates))
 #'
-#'  testit::assert("operator" %in% names(out$operators))
-#'  testit::assert("p" %in% names(out$operators))
-#'  testit::assert("accept" %in% names(out$operators))
-#'  testit::assert("reject" %in% names(out$operators))
-#'  testit::assert("acceptFC" %in% names(out$operators))
-#'  testit::assert("rejectFC" %in% names(out$operators))
-#'  testit::assert("rejectIv" %in% names(out$operators))
-#'  testit::assert("rejectOp" %in% names(out$operators))
+#'  expect_true("operator" %in% names(out$operators))
+#'  expect_true("p" %in% names(out$operators))
+#'  expect_true("accept" %in% names(out$operators))
+#'  expect_true("reject" %in% names(out$operators))
+#'  expect_true("acceptFC" %in% names(out$operators))
+#'  expect_true("rejectFC" %in% names(out$operators))
+#'  expect_true("rejectIv" %in% names(out$operators))
+#'  expect_true("rejectOp" %in% names(out$operators))
 #' @seealso Use \code{\link[tracerer]{remove_burn_ins}}
 #'   to remove the burn-ins from
 #'   the posterior's estimates (\code{posterior$estimates})
