+++ NAMESPACE
@@ -4,4 +4,5 @@ export(bbt_run)
 export(bbt_run_from_model)
 export(get_babette_path)
 export(get_babette_paths)
+export(get_bbt_run_test_output)
 export(plot_densitree)

[…]
