### 2019-03-27 6f6449b
Builds without BEAST2 installed, like CRAN
