### 2019-03-27 d111002
Builds without BEAST2 installed, like CRAN
