+++ vignettes/demo.Rmd
@@ -26,6 +26,11 @@ First, load the library:
 library(babette)
 ```
 
+
+```{r}
+out <- get_bbt_run_test_output() 
+```
+
 The following example shows how to analyze one alignment:
 
  * How to let `babette` run BEAST2
@@ -45,10 +50,12 @@ mcmc <- create_mcmc(
 For a reliable inference, use an ESS of at least 200.
 
 ```{r cache=TRUE}
-out <- bbt_run(
-  get_babette_path("anthus_aco.fas"),
-  mcmc = mcmc
-)
+if (is_beast2_installed()) {
+  out <- bbt_run(
+    get_babette_path("anthus_aco.fas"),
+    mcmc = mcmc
+  )
+}
 ```
 
 ```{r}
