### 2019-03-27 e9dec98
Builds without BEAST2 installed, like CRAN
