+++ R/bbt_run_from_model.R
@@ -41,38 +41,41 @@
 #' }
 #' @author Richèl J.C. Bilderbeek
 #' @examples
-#'  out <- bbt_run_from_model(
-#'    fasta_filename = get_babette_path("anthus_aco.fas"),
-#'    inference_model = create_inference_model(
-#'      mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
-#'    )
-#'  )
+#'   if (is_beast2_installed()) {
 #'
-#'  librar(testthat)
-#'  expect_true("estimates" %in% names(out))
-#'  expect_true("anthus_aco_trees" %in% names(out))
-#'  expect_true("operators" %in% names(out))
-#'  expect_true("output" %in% names(out))
-#'  expect_true(class(out$anthus_aco_trees[[1]]) == "phylo")
-#'  expect_true(length(out$anthus_aco_trees) == 2)
+#'     out <- bbt_run_from_model(
+#'       fasta_filename = get_babette_path("anthus_aco.fas"),
+#'       inference_model = create_inference_model(
+#'         mcmc = create_mcmc(chain_length = 1000, store_every = 1000)
+#'       )
+#'     )
 #'
-#'  expect_true("Sample" %in% names(out$estimates))
-#'  expect_true("posterior" %in% names(out$estimates))
-#'  expect_true("likelihood" %in% names(out$estimates))
-#'  expect_true("prior" %in% names(out$estimates))
-#'  expect_true("treeLikelihood" %in% names(out$estimates))
-#'  expect_true("TreeHeight" %in% names(out$estimates))
-#'  expect_true("YuleModel" %in% names(out$estimates))
-#'  expect_true("birthRate" %in% names(out$estimates))
+#'     library(testthat)
+#'     expect_true("estimates" %in% names(out))
+#'     expect_true("anthus_aco_trees" %in% names(out))
+#'     expect_true("operators" %in% names(out))
+#'     expect_true("output" %in% names(out))
+#'     expect_true(class(out$anthus_aco_trees[[1]]) == "phylo")
+#'     expect_true(length(out$anthus_aco_trees) == 2)
 #'
-#'  expect_true("operator" %in% names(out$operators))
-#'  expect_true("p" %in% names(out$operators))
-#'  expect_true("accept" %in% names(out$operators))
-#'  expect_true("reject" %in% names(out$operators))
-#'  expect_true("acceptFC" %in% names(out$operators))
-#'  expect_true("rejectFC" %in% names(out$operators))
-#'  expect_true("rejectIv" %in% names(out$operators))
-#'  expect_true("rejectOp" %in% names(out$operators))
+#'     expect_true("Sample" %in% names(out$estimates))
+#'     expect_true("posterior" %in% names(out$estimates))
+#'     expect_true("likelihood" %in% names(out$estimates))
+#'     expect_true("prior" %in% names(out$estimates))
+#'     expect_true("treeLikelihood" %in% names(out$estimates))
+#'     expect_true("TreeHeight" %in% names(out$estimates))
+#'     expect_true("YuleModel" %in% names(out$estimates))
+#'     expect_true("birthRate" %in% names(out$estimates))
+#'
+#'     expect_true("operator" %in% names(out$operators))
+#'     expect_true("p" %in% names(out$operators))
+#'     expect_true("accept" %in% names(out$operators))
+#'     expect_true("reject" %in% names(out$operators))
+#'     expect_true("acceptFC" %in% names(out$operators))
+#'     expect_true("rejectFC" %in% names(out$operators))
+#'     expect_true("rejectIv" %in% names(out$operators))
+#'     expect_true("rejectOp" %in% names(out$operators))
+#'   }
[…]
