+++ R/plot_densitree.R
@@ -6,11 +6,13 @@
 #' @return nothing. Will produce a plot.
 #' @author Richèl J.C. Bilderbeek
 #' @examples
-#'   out <- bbt_run(
-#'    get_babette_path("anthus_aco.fas"),
-#'    mcmc = create_mcmc(chain_length = 10000, store_every = 1000)
-#'  )
-#'  plot_densitree(out$anthus_aco_trees)
+#'   if (is_beast2_installed()) {
+#'      out <- bbt_run(
+#'       get_babette_path("anthus_aco.fas"),
+#'       mcmc = create_mcmc(chain_length = 10000, store_every = 1000)
+#'     )
+#'     plot_densitree(out$anthus_aco_trees)
+#'   }
 #' @export
 plot_densitree <- function(
   phylos,
