### 2019-03-27 cda1ba7
Builds without BEAST2 installed, like CRAN
