+++ tests/testthat/test-bbt_check_beast2_packages.R
@@ -2,6 +2,8 @@ context("bbt_check_beast2_packages")
 
 test_that("use", {
 
+  if (!beastier::is_beast2_installed()) return()
+
   # This test uninstalls the NS BEAST2 package.
   # Only do that on CI services, else a user without internet
   # suddenly finds the NS BEAST2 package installed and unable
