### 2019-03-27 25723b6
Thanks @lintr-bot
