### 2019-03-27 07f85c1
Travis installs BEAST2 again
