### 2019-03-30 c11f706
Fix tests
