### 2019-04-09 2a1cb37
Workaround for devtools error, see https://github.com/r-lib/devtools/issues/2020
