+++ .travis.yml
@@ -30,6 +30,8 @@ r_github_packages:
   - MangoTheCat/goodpractice
   - ropensci/spelling
 
+# temporary workaround for failed devtools install, 
+# see https://github.com/r-lib/devtools/issues/2020
 before_install: R -e 'source("https://install-github.me/r-lib/remotes")'
 
 before_script:
