+++ .travis.yml
@@ -11,9 +11,8 @@ matrix:
 
 cache: packages
 
-# No devtools, see https://github.com/r-lib/devtools/issues/2020
 r_packages:
-  # - devtools
+  - devtools
   - ggplot2
   - hunspell
   - magrittr
@@ -29,10 +28,6 @@ r_github_packages:
   - ropensci/mauricer
   - MangoTheCat/goodpractice
   - ropensci/spelling
-
-# temporary workaround for failed devtools install, 
-# see https://github.com/r-lib/devtools/issues/2020
-before_install: R -e 'source("https://install-github.me/r-lib/remotes")'
 
 before_script:
   - # Java
