+++ .travis.yml
@@ -6,8 +6,8 @@ matrix:
     os: linux
   - r: devel
     os: linux
-  - r: release
-    os: osx
+  #- r: release
+  #  os: osx
 
 cache: packages
 
