### 2019-05-02 c0e7859
Remove 'doc' folder
