+++ doc/examples.md
@@ -1,178 +1,0 @@
-# Examples
-
- * Want something more interactive? [lumier](https://github.com/ropensci/lumier) 
-   is an R Shiny app to help create the R function call needed!
-
-For all examples, do load `babette`:
-
-```
-library(babette)
-```
-
-All examples read the alignment from a FASTA file (usually `my_fasta.fas`) 
-and create a BEAST2 input file called `my_beast.xml`.
-
-## Example #1: all default
-
-Using all default settings, only specify a DNA alignment.
-
-![Example #1: all default](all_default.png)
-
-```
-posterior <- bbt_run(
-  "test_output_0.fas"
-)
-```
-
-All other parameters are set to their defaults, as in BEAUti.
-
-## Example #2: fixed crown age
-
-Using all default settings, only specify a DNA alignment.
-
-```
-[No screenshot, as this cannot be done in BEAUti yet]
-```
-
-```
-# Deprecated, use an MRCA prior instead
-posterior <- bbt_run(
-  "my_fasta.fas",
-  posterior_crown_age = 15
-)
-```
-
-An alternative is to date the node of the most recent common ancestor
-of all taxa:
-
-![Example #2: using a MRCA to specify a crown age](mrca_crown_age.png)
-
-```{r example_2_mrca, cache=FALSE}
-posterior <- bbt_run(
-  "my_fasta.fas",
-  mcmc = mcmc,
-  mrca_prior = create_mrca_prior(
-    taxa_names = get_taxa_names("my_fasta.fas"),
-    alignment_id = get_alignment_id("my_fasta.fas"),
-    is_monophyletic = TRUE,
-    mrca_distr = create_normal_distr(
-      mean = 15.0,
-      sigma = 0.025
-    )
-  )
-)
-```
-
-Here we use an MRCA prior with fixed (non-estimated) values of the mean
-and standard deviation for the common ancestor node's time.
-
-
-## Example #3: JC69 site model
-
-![Example #3: JC69 site model](jc69_2_4.png)
-
-```
-posterior <- bbt_run(
-  "my_alignment.fas",
-  site_model = create_jc69_site_model()
-)
-```
-
-## Example #4: Relaxed clock log normal
-
-![Example #4: Relaxed clock log normal](rln_2_4.png)
-
-```{r example_4}
-posterior <- bbt_run(
-  "my_alignment.fas",
-  clock_model = create_rln_clock_model()
-)
-```
-
-## Example #5: Birth-Death tree prior
-
-![Example #5: Birth-Death tree prior](bd_2_4.png)
-
-```{r example_5}
-posterior <- bbt_run(
-  "my_alignment.fas",
-  tree_prior = create_bd_tree_prior() 
-)
-```
-
-## Example #6: Yule tree prior with a normally distributed birth rate
-
-![Example #6: Yule tree prior with a normally distributed birth rate](birth_rate_normal_2_4.png)
-
-```{r example_6}
-posterior <- bbt_run(
-  "my_alignment.fas",
-  tree_prior = create_yule_tree_prior(
-    birth_rate_distr = create_normal_distr()
-  ) 
-)
-```
-
-Thanks to Yacine Ben Chehida for this use case
-
-## Example #7: HKY site model with a non-zero proportion of invariants
-
-![Example #7: HKY site model with a non-zero proportion of invariants](hky_prop_invariant_0_5_2_4.png)
-
-```{r example_7}
-posterior <- bbt_run(
-  "my_alignment.fas",
-  site_model = create_hky_site_model(
-    gamma_site_model = create_gamma_site_model(prop_invariant = 0.5)
-  )
-)
-```
-
-Thanks to Yacine Ben Chehida for this use case
-
[…]
