+++ doc/README.md
@@ -1,9 +1,0 @@
-# `doc`
-
-The `babette` documentation.
-
- * [Code of Conduct](code_of_conduct.md)
- * [Examples](examples.md)
- * [FAQ](faq.md)
- * [Install](install.md)
-

[…]
