### 2019-05-02 bead324
Clean demo vignette
