+++ R/bbt_process_pkg_output.R
@@ -1,4 +1,5 @@
 #' Process the BEAST2 output dependent on BEAST2 package specifics
+#' @inheritParams bbt_default_params_doc
 #' @param out complete babette output
 #' @param mcmc an MCMC
 #' @param alignment_ids alignment IDs, as obtained by
@@ -18,14 +19,15 @@
 bbt_process_pkg_output <- function(
   out,
   mcmc,
+  beast2_working_dir,
   alignment_ids = NA
 ) {
   if (beautier::is_mcmc_nested_sampling(mcmc)) {
     out$ns <- bbt_create_ns(out$output) # nolint internal function
     testit::assert(!beautier::is_one_na(alignment_ids))
     testit::assert(length(alignment_ids) == 1)
-    ns_log_filename <- paste0(alignment_ids, ".posterior.log")
-    ns_trees_filename <- paste0(alignment_ids, ".posterior.trees")
+    ns_log_filename <- file.path(beast2_working_dir, paste0(alignment_ids, ".posterior.log"))
+    ns_trees_filename <- file.path(beast2_working_dir, paste0(alignment_ids, ".posterior.trees"))
     out$ns$estimates <- tracerer::parse_beast_log(ns_log_filename)
     out$ns$trees <- tracerer::parse_beast_trees(ns_trees_filename)
     file.remove(ns_log_filename)
