+++ R/bbt_run_from_model.R
@@ -119,6 +119,7 @@ bbt_run_from_model <- function(
     beast2_output_log_filename = beast2_options$output_log_filename,
     beast2_output_trees_filenames = beast2_options$output_trees_filenames,
     beast2_output_state_filename = beast2_options$output_state_filename,
+    beast2_working_dir = beast2_options$beast2_working_dir,
     beast2_path = beast2_options$beast2_path,
     overwrite = beast2_options$overwrite,
     verbose = beast2_options$verbose
